split ap0202at and ar0147 implemnetations
* creates a public API for controlling the AP0202AT ISP
* employs the API to handle AR0147 specific reset actions
