+++ src/omv/sensors/ap0202at.h
@@ -14,6 +14,32 @@
 // Range 24 to 30 MHz does not require PLL reconfiguration.
 #define AP0202AT_XCLK_FREQ    (24000000)
 
+// register access
+int ap0202at_read_reg_direct(sensor_t *sensor, uint16_t reg_addr, uint16_t* reg_data);
+int ap0202at_write_reg_direct(sensor_t *sensor, uint16_t reg_addr, uint16_t data);
+int ap0202at_write_reg_masked(sensor_t *sensor, uint16_t reg_addr, uint16_t data, uint16_t mask);
+int ap0202at_write_reg_burst_addr_24(sensor_t* sensor, uint16_t *data, uint16_t data_words);
+
+// sensor register passthrough access
+int ap0202at_write_sensor_u16(sensor_t *sensor, uint16_t addr, uint16_t data);
+int ap0202at_write_sensor_sequencer(sensor_t* sensor, uint16_t port_address, uint16_t* sequencer_data, size_t sequencer_words);
+
+// host command interface
+int ap0202at_poll_doorbell_bit(sensor_t *sensor, uint16_t timeout_ms);
+int ap0202at_issue_host_command(sensor_t *sensor, uint16_t command);
+int ap0202at_read_host_command_result(sensor_t *sensor, uint16_t *result);
+int ap0202at_host_command(sensor_t sensor, uint16_t command, uint16_t* result);
+
+// configuration mode
+int ap0202at_enter_configuration_mode_hardware(sensor_t *sensor);
+int ap0202at_enter_configuration_mode_software(sensor_t *sensor);
+
+// reset process
+int ap0202at_reset(sensor_t *sensor);
+int ap0202at_sensor_discovery(sensor_t *sensor, uint16_t *sensor_id);
+int ap0202at_enter_streaming(sensor_t *sensor);
+
+// sensor_t struct initialization
 int ap0202at_init(sensor_t *sensor);
 
 #endif // __AP0202AT_H__
