+++ src/omv/sensors/ap0202at_ar0147.c
@@ -1,0 +1,1114 @@
+/*
+ * This file is part of the OpenMV project.
+ *
+ * Copyright (c) 2023 oclyke <oclyke@oclyke.dev>
+ *
+ * This work is licensed under the MIT license, see the file LICENSE for details.
+ *
+ * AR0147 w/ AP0202AT ISP driver.
+ */
+#include "omv_boardconfig.h"
+#if (OMV_ENABLE_AP0202AT_AR0147 == 1)
+
+#include "openmv/src/omv/sensors/ap0202at.h"
+
+#define AP0147_SENSOR_ID ()
+
+// Forward declarations.
+static const uint16_t patch_28d4_data[];
+static const uint16_t patch_01d4_data[];
+static const uint16_t patch_03d4_data[];
+static const uint16_t patch_11d4_data[];
+static const uint16_t patch_21d4_data[];
+static const uint16_t patch_37d4_data[];
+static const uint16_t patch_39d4_data[];
+static const uint16_t ar0147_sequencer_data[];
+
+/**
+ * @brief Installs the sensor register write workaround.
+ * @ref AP0202AT-REV2.ini lines 331, 370
+ * 
+ * @param sensor the sensor struct.
+ * @return int 0 on success, -1 on error.
+ */
+static int install_sensor_reg_write_workaround(sensor_t *sensor) {
+    int ret = 0;
+
+    // This workaround is only applicable to the
+    //   AR0147R3 SE+T2 sensor. If the sensor is not the
+    //   AR0147R3 SE+T2, then return success.
+    if (false) {
+        return 0;
+    }
+
+    // AP0202AT-REV2.ini line 372
+    static uint16_t reg_data0[] = {
+        0x0016, 0x6039, // Enable clock to sensor
+        0x0032, 0xC0F0, // Bring sensor out of reset
+        0x3B00, 0x2083,
+        0x3B02, 0x3500,
+        0x3B04, 0x0100,
+        0x3B88, 0x0005,
+    }
+
+    // AP0202AT-REV2.ini line 379
+    static uint16_t reg_data1[] = {
+        0x3B00, 0x2083,
+        0x3B02, 0x3006,
+        0x3B04, 0x0008,
+        0x3B88, 0x0005,
+        0x3B00, 0x2083,
+        0x3B02, 0x3F66,
+        0x3B04, 0x00C0,
+        0x3B88, 0x0005,
+    }
+
+    // Loop over the first set of register writes.
+    for (uint16_t i = 0; i < sizeof(reg_data0); i += 2) {
+        if (ap0202at_write_reg_direct(sensor, reg_data0[i], reg_data0[i + 1]) != 0) {
+            return -1;
+        }
+    }
+
+    // AP0202AT-REV2.ini line 378
+    delay_ms(1);
+
+    // Loop over the second set of register writes.
+    for (uint16_t i = 0; i < sizeof(reg_data1); i += 2) {
+        if (ap0202at_write_reg_direct(sensor, reg_data1[i], reg_data1[i + 1]) != 0) {
+            return -1;
+        }
+    }
+
+    // AP0202AT-REV2.ini line 387
+    delay_ms(50);
+
+    return ret;
+}
+
+/**
+ * @brief Loads and applies patch 28d4.
+ * 
+ * APA0202AT-REV2_AR0147-REV3.ini line 1108
+ * 
+ * @param sensor the sensor struct.
+ * @return int 0 on success, -1 on error.
+ */
+static int load_apply_patch_28d4(sensor_t *sensor) {
+    int ret = 0;
+    uint16_t host_command_result;
+
+    // APA0202AT-REV2_AR0147-REV3.ini line 1114
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_VAR_CMD_HANDLER_PARAMS_POOL_0, 0x5d8) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_VAR_CMD_HANDLER_PARAMS_POOL_1, 0x2364) != 0) {
+        return -1;
+    }
+    if  (ap0202at_host_command(sensor, 0x8706, &host_command_result) != 0) {
+        return -1;
+    }
+    if (host_command_result != AP0202AT_HC_RESP_ENOERR) {
+        return -1;
+    }
+
+    // APA0202AT-REV2_AR0147-REV3.ini line 1121
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_REG_ACCESS_CTL_STAT, 0x0001) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_REG_PHYSICAL_ADDRESS_ACCESS, 0x4d28) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_burst_addr_24(sensor, patch_28d4_data, sizeof(patch_28d4_data) / sizeof(uint16_t)) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_REG_LOGICAL_ADDRESS_ACCESS, 0x0000) != 0) {
+        return -1;
+    }
+
+    // APA0202AT-REV2_AR0147-REV3.ini line 1316
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_VAR_CMD_HANDLER_PARAMS_POOL_0, 0x2504) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_VAR_CMD_HANDLER_PARAMS_POOL_1, 0x28d4) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_VAR_CMD_HANDLER_PARAMS_POOL_2, 0xa103) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_VAR_CMD_HANDLER_PARAMS_POOL_3, 0x0204) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_VAR_CMD_HANDLER_PARAMS_POOL_4, 0x2364) != 0) {
+        return -1;
+    }
+    if  (ap0202at_host_command(sensor, 0x8702, &host_command_result) != 0) {
+        return -1;
+    }
+    if (host_command_result != AP0202AT_HC_RESP_ENOERR) {
+        return -1;
+    }
+
+    // APA0202AT-REV2_AR0147-REV3.ini line 1327
+    if  (ap0202at_host_command(sensor, 0x8701, &host_command_result) != 0) {
+        return -1;
+    }
+    if (host_command_result != AP0202AT_HC_RESP_ENOERR) {
+        return -1;
+    }
+
+    return ret;
+}
+
+/**
+ * @brief Loads and applies patch 01d4.
+ * 
+ * APA0202AT-REV2_AR0147-REV3.ini line 822
+ * 
+ * @param sensor the sensor struct.
+ * @return int 0 on success, -1 on error.
+ */
+static int load_apply_patch_01d4(sensor_t *sensor) {
+    int ret = 0;
+    uint16_t host_command_result;
+
+    // APA0202AT-REV2_AR0147-REV3.ini line 829
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_VAR_CMD_HANDLER_PARAMS_POOL_0, 0x0) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_VAR_CMD_HANDLER_PARAMS_POOL_1, 0x54) != 0) {
+        return -1;
+    }
+    if  (ap0202at_host_command(sensor, 0x8706, &host_command_result) != 0) {
+        return -1;
+    }
+    if (host_command_result != AP0202AT_HC_RESP_ENOERR) {
+        return -1;
+    }
+
+    // APA0202AT-REV2_AR0147-REV3.ini line 836
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_REG_ACCESS_CTL_STAT, 0x0001) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_REG_PHYSICAL_ADDRESS_ACCESS, 0x4750) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_burst_addr_24(sensor, patch_01d4_data, sizeof(patch_01d4_data) / sizeof(uint16_t)) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_REG_LOGICAL_ADDRESS_ACCESS, 0x0000) != 0) {
+        return -1;
+    }
+
+    // APA0202AT-REV2_AR0147-REV3.ini line 843
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_VAR_CMD_HANDLER_PARAMS_POOL_0, 0x0030) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_VAR_CMD_HANDLER_PARAMS_POOL_1, 0x01d4) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_VAR_CMD_HANDLER_PARAMS_POOL_2, 0xa103) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_VAR_CMD_HANDLER_PARAMS_POOL_3, 0x0204) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_VAR_CMD_HANDLER_PARAMS_POOL_4, 0x0054) != 0) {
+        return -1;
+    }
+    if  (ap0202at_host_command(sensor, 0x8702, &host_command_result) != 0) {
+        return -1;
+    }
+    if (host_command_result != AP0202AT_HC_RESP_ENOERR) {
+        return -1;
+    }
+
+    // APA0202AT-REV2_AR0147-REV3.ini line 854
+    if  (ap0202at_host_command(sensor, 0x8701, &host_command_result) != 0) {
+        return -1;
+    }
+    if (host_command_result != AP0202AT_HC_RESP_ENOERR) {
+        return -1;
+    }
+
+    return ret;
+}
+
+/**
+ * @brief Loads and applies patch 03d4.
+ * 
+ * APA0202AT-REV2_AR0147-REV3.ini line 860
+ * 
+ * @param sensor the sensor struct.
+ * @return int 0 on success, -1 on error.
+ */
+static int load_apply_patch_03d4(sensor_t *sensor) {
+    int ret = 0;
+    uint16_t host_command_result;
+
+    // APA0202AT-REV2_AR0147-REV3.ini line 867
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_VAR_CMD_HANDLER_PARAMS_POOL_0, 0x54) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_VAR_CMD_HANDLER_PARAMS_POOL_1, 0x98) != 0) {
+        return -1;
+    }
+    if  (ap0202at_host_command(sensor, 0x8706, &host_command_result) != 0) {
+        return -1;
+    }
+    if (host_command_result != AP0202AT_HC_RESP_ENOERR) {
+        return -1;
+    }
+
+    // APA0202AT-REV2_AR0147-REV3.ini line 874
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_REG_ACCESS_CTL_STAT, 0x0001) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_REG_PHYSICAL_ADDRESS_ACCESS, 0x47a4) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_burst_addr_24(sensor, patch_03d4_data, sizeof(patch_03d4_data) / sizeof(uint16_t)) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_REG_LOGICAL_ADDRESS_ACCESS, 0x0000) != 0) {
+        return -1;
+    }
+
+    // APA0202AT-REV2_AR0147-REV3.ini line 884
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_VAR_CMD_HANDLER_PARAMS_POOL_0, 0x00c8) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_VAR_CMD_HANDLER_PARAMS_POOL_1, 0x03d4) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_VAR_CMD_HANDLER_PARAMS_POOL_2, 0xa103) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_VAR_CMD_HANDLER_PARAMS_POOL_3, 0x0204) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_VAR_CMD_HANDLER_PARAMS_POOL_4, 0x0098) != 0) {
+        return -1;
+    }
+    if  (ap0202at_host_command(sensor, 0x8702, &host_command_result) != 0) {
+        return -1;
+    }
+    if (host_command_result != AP0202AT_HC_RESP_ENOERR) {
+        return -1;
+    }
+
+    // APA0202AT-REV2_AR0147-REV3.ini line 894
+    if  (ap0202at_host_command(sensor, 0x8701, &host_command_result) != 0) {
+        return -1;
+    }
+    if (host_command_result != AP0202AT_HC_RESP_ENOERR) {
+        return -1;
+    }
+
+    return ret;
+}
+
+
+/**
+ * @brief Loads and applies patch 11d4.
+ * 
+ * APA0202AT-REV2_AR0147-REV3.ini line 900
+ * 
+ * @param sensor the sensor struct.
+ * @return int 0 on success, -1 on error.
+ */
+static int load_apply_patch_11d4(sensor_t *sensor) {
+    int ret = 0;
+    uint16_t host_command_result;
+
+    // APA0202AT-REV2_AR0147-REV3.ini line 907
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_VAR_CMD_HANDLER_PARAMS_POOL_0, 0x00ec) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_VAR_CMD_HANDLER_PARAMS_POOL_1, 0x0118) != 0) {
+        return -1;
+    }
+    if  (ap0202at_host_command(sensor, 0x8706, &host_command_result) != 0) {
+        return -1;
+    }
+    if (host_command_result != AP0202AT_HC_RESP_ENOERR) {
+        return -1;
+    }
+
+    // APA0202AT-REV2_AR0147-REV3.ini line 914
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_REG_ACCESS_CTL_STAT, 0x0001) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_REG_PHYSICAL_ADDRESS_ACCESS, 0x483c) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_burst_addr_24(sensor, patch_11d4_data, sizeof(patch_11d4_data) / sizeof(uint16_t)) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_REG_LOGICAL_ADDRESS_ACCESS, 0x0000) != 0) {
+        return -1;
+    }
+
+    // APA0202AT-REV2_AR0147-REV3.ini line 925
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_VAR_CMD_HANDLER_PARAMS_POOL_0, 0x019c) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_VAR_CMD_HANDLER_PARAMS_POOL_1, 0x11d4) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_VAR_CMD_HANDLER_PARAMS_POOL_2, 0xa103) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_VAR_CMD_HANDLER_PARAMS_POOL_3, 0x0204) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_VAR_CMD_HANDLER_PARAMS_POOL_4, 0x0118) != 0) {
+        return -1;
+    }
+    if  (ap0202at_host_command(sensor, 0x8702, &host_command_result) != 0) {
+        return -1;
+    }
+    if (host_command_result != AP0202AT_HC_RESP_ENOERR) {
+        return -1;
+    }
+
+    // APA0202AT-REV2_AR0147-REV3.ini line 936
+    if  (ap0202at_host_command(sensor, 0x8701, &host_command_result) != 0) {
+        return -1;
+    }
+    if (host_command_result != AP0202AT_HC_RESP_ENOERR) {
+        return -1;
+    }
+
+    return ret;
+}
+
+
+/**
+ * @brief Loads and applies patch 21d4.
+ * 
+ * APA0202AT-REV2_AR0147-REV3.ini line 900
+ * 
+ * @param sensor the sensor struct.
+ * @return int 0 on success, -1 on error.
+ */
+static int load_apply_patch_21d4(sensor_t *sensor) {
+    int ret = 0;
+    uint16_t host_command_result;
+
+    // APA0202AT-REV2_AR0147-REV3.ini line 1036
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_VAR_CMD_HANDLER_PARAMS_POOL_0, 0x4e0) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_VAR_CMD_HANDLER_PARAMS_POOL_1, 0xb8) != 0) {
+        return -1;
+    }
+    if  (ap0202at_host_command(sensor, 0x8706, &host_command_result) != 0) {
+        return -1;
+    }
+    if (host_command_result != AP0202AT_HC_RESP_ENOERR) {
+        return -1;
+    }
+
+    // APA0202AT-REV2_AR0147-REV3.ini line 1044
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_REG_ACCESS_CTL_STAT, 0x0001) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_REG_PHYSICAL_ADDRESS_ACCESS, 0x4c30) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_burst_addr_24(sensor, patch_21d4_data, sizeof(patch_21d4_data) / sizeof(uint16_t)) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_REG_LOGICAL_ADDRESS_ACCESS, 0x0000) != 0) {
+        return -1;
+    }
+
+    // APA0202AT-REV2_AR0147-REV3.ini line 1054
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_VAR_CMD_HANDLER_PARAMS_POOL_0, 0x056c) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_VAR_CMD_HANDLER_PARAMS_POOL_1, 0x21d4) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_VAR_CMD_HANDLER_PARAMS_POOL_2, 0xa103) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_VAR_CMD_HANDLER_PARAMS_POOL_3, 0x0204) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_VAR_CMD_HANDLER_PARAMS_POOL_4, 0x00b8) != 0) {
+        return -1;
+    }
+    if  (ap0202at_host_command(sensor, 0x8702, &host_command_result) != 0) {
+        return -1;
+    }
+    if (host_command_result != AP0202AT_HC_RESP_ENOERR) {
+        return -1;
+    }
+
+    // APA0202AT-REV2_AR0147-REV3.ini line 1064
+    if  (ap0202at_host_command(sensor, 0x8701, &host_command_result) != 0) {
+        return -1;
+    }
+    if (host_command_result != AP0202AT_HC_RESP_ENOERR) {
+        return -1;
+    }
+
+    return ret;
+}
+
+/**
+ * @brief Loads and applies patch 37d4.
+ * 
+ * APA0202AT-REV2_AR0147-REV3.ini line 1432
+ * 
+ * @param sensor the sensor struct.
+ * @return int 0 on success, -1 on error.
+ */
+static int load_apply_patch_37d4(sensor_t *sensor) {
+    int ret = 0;
+    uint16_t host_command_result;
+
+    // APA0202AT-REV2_AR0147-REV3.ini line 1434
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_VAR_CMD_HANDLER_PARAMS_POOL_0, 0x293c) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_VAR_CMD_HANDLER_PARAMS_POOL_1, 0x218) != 0) {
+        return -1;
+    }
+    if  (ap0202at_host_command(sensor, 0x8706, &host_command_result) != 0) {
+        return -1;
+    }
+    if (host_command_result != AP0202AT_HC_RESP_ENOERR) {
+        return -1;
+    }
+
+    // APA0202AT-REV2_AR0147-REV3.ini line 1441
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_REG_ACCESS_CTL_STAT, 0x0001) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_REG_PHYSICAL_ADDRESS_ACCESS, 0x708c) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_burst_addr_24(sensor, patch_37d4_data, sizeof(patch_37d4_data) / sizeof(uint16_t)) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_REG_LOGICAL_ADDRESS_ACCESS, 0x0000) != 0) {
+        return -1;
+    }
+
+    // APA0202AT-REV2_AR0147-REV3.ini line 1459
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_VAR_CMD_HANDLER_PARAMS_POOL_0, 0x2b30) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_VAR_CMD_HANDLER_PARAMS_POOL_1, 0x37d4) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_VAR_CMD_HANDLER_PARAMS_POOL_2, 0xa103) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_VAR_CMD_HANDLER_PARAMS_POOL_3, 0x0204) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_VAR_CMD_HANDLER_PARAMS_POOL_4, 0x0218) != 0) {
+        return -1;
+    }
+    if  (ap0202at_host_command(sensor, 0x8702, &host_command_result) != 0) {
+        return -1;
+    }
+    if (host_command_result != AP0202AT_HC_RESP_ENOERR) {
+        return -1;
+    }
+
+    // APA0202AT-REV2_AR0147-REV3.ini line 1469
+    if  (ap0202at_host_command(sensor, 0x8701, &host_command_result) != 0) {
+        return -1;
+    }
+    if (host_command_result != AP0202AT_HC_RESP_ENOERR) {
+        return -1;
+    }
+
+    return ret;
+}
+
+/**
+ * @brief Loads and applies patch 39d4.
+ * 
+ * APA0202AT-REV2_AR0147-REV3.ini line 
+ * 
+ * @param sensor the sensor struct.
+ * @return int 0 on success, -1 on error.
+ */
+static int load_apply_patch_39d4(sensor_t *sensor) {
+    int ret = 0;
+    uint16_t host_command_result;
+
+    // APA0202AT-REV2_AR0147-REV3.ini line
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_VAR_CMD_HANDLER_PARAMS_POOL_0, 0x2b54) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_VAR_CMD_HANDLER_PARAMS_POOL_1, 0xac) != 0) {
+        return -1;
+    }
+    if  (ap0202at_host_command(sensor, 0x8706, &host_command_result) != 0) {
+        return -1;
+    }
+    if (host_command_result != AP0202AT_HC_RESP_ENOERR) {
+        return -1;
+    }
+
+    // APA0202AT-REV2_AR0147-REV3.ini line
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_REG_ACCESS_CTL_STAT, 0x0001) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_REG_PHYSICAL_ADDRESS_ACCESS, 0x72a4) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_burst_addr_24(sensor, patch_39d4_data, sizeof(patch_39d4_data) / sizeof(uint16_t)) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_REG_LOGICAL_ADDRESS_ACCESS, 0x0000) != 0) {
+        return -1;
+    }
+
+    // APA0202AT-REV2_AR0147-REV3.ini line
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_VAR_CMD_HANDLER_PARAMS_POOL_0, 0x2bd8) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_VAR_CMD_HANDLER_PARAMS_POOL_1, 0x39d4) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_VAR_CMD_HANDLER_PARAMS_POOL_2, 0xa103) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_VAR_CMD_HANDLER_PARAMS_POOL_3, 0x0204) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_reg_direct(sensor, AP0202AT_VAR_CMD_HANDLER_PARAMS_POOL_4, 0x00ac) != 0) {
+        return -1;
+    }
+    if  (ap0202at_host_command(sensor, 0x8702, &host_command_result) != 0) {
+        return -1;
+    }
+    if (host_command_result != AP0202AT_HC_RESP_ENOERR) {
+        return -1;
+    }
+
+    // APA0202AT-REV2_AR0147-REV3.ini line 
+    if  (ap0202at_host_command(sensor, 0x8701, &host_command_result) != 0) {
+        return -1;
+    }
+    if (host_command_result != AP0202AT_HC_RESP_ENOERR) {
+        return -1;
+    }
+
+    return ret;
+}
+
+/**
+ * @brief Loads all applicable patches for the combination
+ *   of the AP0202AT ISP and the image sensor.
+ * 
+ * @param sensor the sensor struct.
+ * @return int 0 on success, -1 on error.
+ */
+static int load_patches(sensor_t *sensor) {
+    int ret = 0;
+
+    if (load_apply_patch_28d4(sensor) != 0) {
+        return -1;
+    }
+    if (load_apply_patch_01d4(sensor) != 0) {
+        return -1;
+    }
+    if (load_apply_patch_03d4(sensor) != 0) {
+        return -1;
+    }
+    if (load_apply_patch_11d4(sensor) != 0) {
+        return -1;
+    }
+    if (load_apply_patch_21d4(sensor) != 0) {
+        return -1;
+    }
+    if (load_apply_patch_37d4(sensor) != 0) {
+        return -1;
+    }
+    if (load_apply_patch_39d4(sensor) != 0) {
+        return -1;
+    }
+
+    return ret;
+}
+
+/**
+ * @brief Reset the image sensor into the default state.
+ * 
+ * @param sensor the sensor struct.
+ * @return int 0 on success, -1 on error.
+ */
+static int reset_image_sensor(sensor_t *sensor) {
+    int ret = 0;
+    uint16_t host_command_result;
+
+    // The only image sensor supported is the AR0147.
+    if (false) {
+        return -1;
+    }
+
+    if (ap0202at_write_reg_masked(
+            sensor,
+            AP0202AT_VAR_SENSOR_MGR_MODE,
+            AP0202AT_VAR_SENSOR_MGR_SENSOR_DEFAULT_SEQUENCER_LOAD_INHIBIT_TRUE,
+            AP0202AT_VAR_SENSOR_MGR_SENSOR_DEFAULT_SEQUENCER_LOAD_INHIBIT_MASK
+        ) != 0) {
+        return -1;
+    }
+
+    // APA0202AT-REV2_AR0147-REV3.ini line 749
+    if  (ap0202at_host_command(sensor, AP0202AT_HC_CMD_CCIMGR_GET_LOCK, &host_command_result) != 0) {
+        return -1;
+    }
+    if (host_command_result != AP0202AT_HC_RESP_ENOERR) {
+        return -1;
+    }
+
+    // APA0202AT-REV2_AR0147-REV3.ini line 750
+    if  (ap0202at_host_command(sensor, AP0202AT_HC_CMD_CCIMGR_LOCK_STATUS, &host_command_result) != 0) {
+        return -1;
+    }
+    if (host_command_result != AP0202AT_HC_CMD_CCIMGR_LOCK_STATUS_RESP_LOCKED) {
+        return -1;
+    }
+
+    // APA0202AT-REV2_AR0147-REV3.ini line 754
+    if (ap0202at_write_sensor_u16(sensor, 0x2512,0x8000) != 0) {
+        return -1;
+    }
+
+    // APA0202AT-REV2_AR0147-REV3.ini line 803
+    if (ap0202at_write_sensor_sequencer(sensor, 0x2510, ar0147_sequencer_data, sizeof(ar0147_sequencer_data)/ sizeof(uint16_t)) != 0) {
+        return -1;
+    }
+
+    // APA0202AT-REV2_AR0147-REV3.ini line 804
+    if  (ap0202at_host_command(sensor, AP0202AT_HC_CMD_CCIMGR_RELEASE_LOCK, &host_command_result) != 0) {
+        return -1;
+    }
+    if (host_command_result != AP0202AT_HC_RESP_ENOERR) {
+        return -1;
+    }
+
+    // APA0202AT-REV2.ini line 338
+    if  (ap0202at_host_command(sensor, AP0202AT_HC_CMD_SENSOR_MGR_INITIALIZE_SENSOR, &host_command_result) != 0) {
+        return -1;
+    }
+    if (host_command_result != AP0202AT_HC_RESP_ENOERR) {
+        return -1;
+    }
+
+    // APA0202AT-REV2_AR0147-REV3.ini line 733
+    if  (ap0202at_host_command(sensor, AP0202AT_HC_CMD_CCIMGR_GET_LOCK, &host_command_result) != 0) {
+        return -1;
+    }
+    if (host_command_result != AP0202AT_HC_RESP_ENOERR) {
+        return -1;
+    }
+
+    // APA0202AT-REV2_AR0147-REV3.ini line 738
+    if (ap0202at_write_sensor_u16(sensor, 0x3754, 0x0000) != 0) {
+        return -1;
+    }
+    if (ap0202at_write_sensor_u16(sensor, 0x3F90, 0x0800) != 0) {
+        return -1;
+    }
+
+    // APA0202AT-REV2_AR0147-REV3.ini line 741
+    if  (ap0202at_host_command(sensor, AP0202AT_HC_CMD_CCIMGR_RELEASE_LOCK, &host_command_result) != 0) {
+        return -1;
+    }
+    if (host_command_result != AP0202AT_HC_RESP_ENOERR) {
+        return -1;
+    }
+
+
+    return ret;
+}
+
+/**
+ * @brief Reset the AP0202AT ISP and the AR0147 into the
+ *   default state.
+ * 
+ * - 
+ * 
+ * @param sensor Pointer to sensor struct.
+ * @return int 0 on success, -1 on error.
+ */
+static int reset(sensor_t *sensor) {
+    int ret = 0;
+    uint16_t sensor_id;
+
+    // Reset the AP0202AT
+    ret = ap0202at_reset(sensor);
+    if (ret != 0) {
+        return -1;
+    }
+
+    // Install the sensor reg write workaround.
+    if (install_sensor_reg_write_workaround(sensor) != 0) {
+        return -1;
+    }
+
+    // Load patches for the AP0202AT ISP.
+    if(load_patches(sensor) != 0) {
+        return -1;
+    }
+
+    // Perform sensor discovery and fail if no sensor is
+    //   found.
+    ret = ap0202at_sensor_discovery(sensor, &sensor_id);
+    if (ret != 0) {
+        return -1;
+    }
+    if (sensor_id != AP0147_SENSOR_ID) {
+        return -1;
+    }
+
+
+    // Perform sensor configuration.
+    if(reset_image_sensor(sensor) != 0) {
+        return -1;
+    }
+
+    // Issue the Change Config command to enter streaming.
+    ret = ap0202at_enter_streaming(sensor);
+    if (ret != 0) {
+        return -1;
+    }
+
+    return ret;
+}
+
+/**
+ * @brief Initialize the sensor structure for the AP0202AT.
+ * 
+ * @param sensor the sensor struct.
+ * @return int 0 on success, -1 on error.
+ */
+int ap0202at_ar0147_init(sensor_t *sensor) {
+    int ret = 0;
+
+    // Initialize the AP0202AT portions of the sensor
+    //   structure.
+    ret = ap0202at_init(sensor);
+    if (ret != 0) {
+        return -1;
+    }
+
+    sensor->reset = reset;
+
+    return 0;
+}
+
+
+// AP0202AT-REV2_AR0147-REV3.ini line 1124
+static const uint16_t patch_28d4_data[] = {
+    0xcd28, 0xc0f1, 0x70cf, 0xffff, 0xb190, 0x9003, 0x0fda, 0x1320, 0xd91e, 0x6901, 0x7810, 0xc0d1, 0x7ee0, 0x72cf, 0xffff, 0xeffc, 0x8a00, 0xe080, 0x208a, 0x0e19, 0x79c0, 0x20ca, 0x0f81, 0x0000, 0x0677,
+    0xcd58, 0x21ca, 0x0061, 0x7fe0, 0xaa20, 0xc0f1, 0x0ebe, 0x1300, 0x705a, 0x77cf, 0xffff, 0xecc0, 0x9700, 0x240a, 0x2f80, 0xffff, 0xec80, 0x0809, 0x0052, 0xe004, 0x7810, 0x204e, 0x0301, 0x8f16, 0x232f,
+    0xcd88, 0x2047, 0xe81b, 0xdd01, 0xf004, 0xe501, 0x7daf, 0x0d6f, 0x1113, 0x9709, 0x143c, 0x2101, 0x2202, 0x234e, 0x6038, 0xe001, 0xe101, 0x0f62, 0x1320, 0x711a, 0x2602, 0x14c0, 0x6038, 0x0f56, 0x1320,
+    0xcdb8, 0x710a, 0xe9ec, 0xf021, 0x8f02, 0x254a, 0x2040, 0xe081, 0x25c2, 0x254c, 0xdd01, 0xf003, 0xe501, 0x7daf, 0x0d33, 0x1113, 0x2202, 0x234e, 0x2602, 0x14c0, 0x2017, 0x0550, 0x143c, 0x2100, 0x2040,
+    0xcde8, 0x0051, 0x9708, 0x0f1e, 0x1320, 0x712a, 0x2002, 0x2040, 0x0f16, 0x1320, 0x712a, 0xe9e9, 0x222f, 0x2388, 0x0671, 0x1320, 0x704a, 0xc0f1, 0x0e32, 0x1300, 0x1024, 0x0080, 0x75cf, 0xffff, 0xecc0,
+    0xce18, 0x7628, 0xe88e, 0xe601, 0x78d0, 0xb5c2, 0xd900, 0xb523, 0xda20, 0x71cf, 0xffff, 0xb5f0, 0xb14b, 0xb14c, 0xf012, 0x70c9, 0x0ed6, 0x1320, 0xd950, 0xb503, 0x9523, 0x8d18, 0x2008, 0x0040, 0xb503,
+    0xce48, 0x9523, 0x2602, 0x1040, 0x7810, 0xffc4, 0xb502, 0x71cf, 0xffff, 0xb5b0, 0x0651, 0x1320, 0xb118, 0x72cf, 0xffff, 0xa168, 0xc0f1, 0xe985, 0xd907, 0xa221, 0xd900, 0xb225, 0xe88f, 0x8221, 0x091b,
+    0xce78, 0x01d1, 0x9225, 0x0913, 0x00df, 0x2240, 0x0281, 0xd808, 0x09ae, 0x04e0, 0xb88d, 0xd809, 0xf158, 0xc0f1, 0xc5e1, 0xc1a8, 0x75cf, 0xffff, 0xa168, 0xda00, 0xe985, 0xd92c, 0xa521, 0xb545, 0x088b,
+    0xcea8, 0x0011, 0x8521, 0xe1ac, 0x0082, 0x000b, 0xf60b, 0x70cf, 0x0000, 0x30b0, 0x097a, 0x04e0, 0x2540, 0x1281, 0xd82d, 0xf033, 0x0971, 0x0b50, 0x0967, 0x0b91, 0x71cf, 0xffff, 0xad34, 0x910b, 0x1c00,
+    0xced8, 0x3004, 0x1c02, 0x3004, 0x1c04, 0x3004, 0x1c06, 0x3004, 0x910c, 0x1c08, 0x3004, 0x1c0a, 0x3004, 0x1c0c, 0x3004, 0x1c0e, 0x3004, 0x910d, 0x1c10, 0x3004, 0x1c12, 0x3004, 0x1c14, 0x3004, 0x1c16,
+    0xcf08, 0x3004, 0x9111, 0x718b, 0x1c18, 0x3004, 0x1c1a, 0x3004, 0x1c1c, 0x3004, 0x1c1e, 0x3004, 0xd865, 0xb807, 0x097a, 0x04e0, 0xda20, 0xd800, 0xa501, 0xd809, 0x0585, 0x1320, 0xc0a8, 0x70cf, 0xffff,
+    0xcf38, 0xad34, 0x9002, 0x2004, 0x0f80, 0x0000, 0x2000, 0x2842, 0x837e, 0x9505, 0xf206, 0xb8a1, 0x7910, 0xb505, 0xf005, 0x204f, 0x0041, 0xb525, 0x70cf, 0x0000, 0x30b0, 0x0886, 0x04c0, 0xd82e, 0xf1e3,
+    0xcf68, 0xc0f1, 0xc5e1, 0x75cf, 0xffff, 0xa168, 0xda00, 0xe984, 0xd909, 0xa521, 0xb545, 0xe8a7, 0x8521, 0xe925, 0x0923, 0x0250, 0x0947, 0x0291, 0x70cf, 0xffff, 0xad34, 0x9002, 0x9525, 0x2186, 0x0006,
+    0xcf98, 0x081d, 0x031e, 0x208a, 0x0006, 0xf00b, 0x70cf, 0x0000, 0x3064, 0x088a, 0x04e0, 0x2540, 0x1281, 0xd80a, 0xf00b, 0xd800, 0x7905, 0x70cf, 0x0000, 0x3064, 0x082a, 0x04e0, 0x7930, 0xd800, 0xa501,
+    0xcfc8, 0xd809, 0x04e9, 0x1300, 0x78e0, 0xc0f1, 0xc5e1, 0x75cf, 0xffff, 0xa168, 0xe984, 0xd90b, 0xa521, 0xe8a9, 0x8521, 0xe927, 0x093b, 0x02d0, 0x094b, 0x0311, 0x70cf, 0xffff, 0xad74, 0x9026, 0x70cf,
+    0xcff8, 0x0000, 0x0fff, 0x2109, 0x0001, 0x9505, 0x2004, 0x0f80, 0xffff, 0xf000, 0xb505, 0x7905, 0x70cf, 0x0000, 0x301e, 0x0fd2, 0x04a0, 0x7930, 0xd800, 0xf009, 0x70cf, 0x0000, 0x301e, 0x080e, 0x04e0,
+    0xd028, 0x2540, 0x1281, 0xd80c, 0xa501, 0xd809, 0x0481, 0x1300, 0x78e0, 0xc0f1, 0xc5e1, 0xe986, 0xdb36, 0x72cf, 0xffff, 0xeffc, 0xaa61, 0xe89b, 0x75cf, 0xffff, 0xeffc, 0x8d41, 0xea17, 0x0a2b, 0x0d91,
+    0xd058, 0x0992, 0x0d00, 0x0821, 0x0250, 0x71cf, 0xffff, 0xecc0, 0x8902, 0x8923, 0xb808, 0x7905, 0x70cf, 0x0000, 0x3516, 0x0f72, 0x0480, 0xd800, 0xad01, 0xd809, 0x0435, 0x1300, 0x78e0, 0xc0f1, 0x0bb6,
+    0xd088, 0x1300, 0x75cf, 0xffff, 0xa168, 0x76cf, 0xffff, 0x8284, 0xe986, 0xd901, 0xa521, 0xd900, 0xb525, 0x0883, 0x0011, 0x8541, 0x0a7f, 0x01d5, 0x2740, 0x7281, 0x7955, 0x7900, 0x0073, 0x0000, 0x001b,
+    0xd0b8, 0x0000, 0x0029, 0x0000, 0x0053, 0x0000, 0x0067, 0x0000, 0x0077, 0x0000, 0x0085, 0x0000, 0x70cf, 0x0000, 0x30ce, 0x0f5e, 0x04a0, 0x2540, 0x1281, 0xd802, 0xf021, 0x0b6e, 0x0c20, 0x9505, 0x7108,
+    0xd0e8, 0xb505, 0x9613, 0x080d, 0x00d1, 0x1620, 0x1080, 0x080d, 0x00d1, 0x214f, 0x0140, 0x7910, 0xb505, 0x70cf, 0x0000, 0x30ce, 0x0ee2, 0x0480, 0xd803, 0xf009, 0x70cf, 0x0000, 0x340a, 0x0f1e, 0x04a0,
+    0xd118, 0x2540, 0x1281, 0xd804, 0xa501, 0xd809, 0x0389, 0x1300, 0x9525, 0x70cf, 0x0000, 0x340a, 0x0eba, 0x04a0, 0x2186, 0x0202, 0xd805, 0xf1f3, 0x70cf, 0x0000, 0x340c, 0x0ef2, 0x04a0, 0x2540, 0x1281,
+    0xd148, 0xd806, 0xf1eb, 0x9525, 0x70cf, 0x0000, 0x340c, 0x0e92, 0x04a0, 0xb987, 0xd800, 0xf1e1, 0x78e0, 0xc0f1, 0xc5e1, 0x75cf, 0xffff, 0xa168, 0xe986, 0xd91c, 0xa521, 0xd900, 0xb525, 0x08af, 0x0011,
+    0xd178, 0x8521, 0xe19c, 0xf22c, 0xf61d, 0x09a3, 0x0011, 0x71cf, 0xffff, 0xaff4, 0x9122, 0x72cf, 0xffff, 0xb530, 0x098d, 0x001e, 0x73cf, 0xffff, 0xb5f0, 0x93a4, 0x9325, 0x0d09, 0x1040, 0x0d0d, 0x1043,
+    0xd1a8, 0x9327, 0xb324, 0x9328, 0xb325, 0x9220, 0xb989, 0xf036, 0x0937, 0x0750, 0x096b, 0x0791, 0x70cf, 0x0000, 0x30c6, 0x71cf, 0xffff, 0xb5f8, 0x0e9e, 0x04a0, 0xda04, 0xd800, 0xf00a, 0x70cf, 0x0000,
+    0xd1d8, 0x30b4, 0x0e5a, 0x04a0, 0x2540, 0x1281, 0xd81d, 0xa501, 0xd809, 0xf01d, 0x70cf, 0xffff, 0xaff4, 0x9022, 0x9505, 0x090f, 0x001e, 0x2045, 0x01c1, 0xb525, 0xd81e, 0xf006, 0x2046, 0x01c0, 0x7910,
+    0xd208, 0xb505, 0xd800, 0xa501, 0x70cf, 0x0000, 0x30b4, 0x0dd2, 0x0480, 0xd809, 0xf005, 0x9220, 0xb9a9, 0xb220, 0x0291, 0x1300, 0x78e0, 0xc0f1, 0xc5e1, 0x75cf, 0xffff, 0xa168, 0xc1a2, 0x08e7, 0x0011,
+    0xd238, 0xe985, 0xd800, 0xb504, 0xd816, 0xa501, 0x8501, 0xe097, 0xf244, 0xf616, 0x08cd, 0x0010, 0x08cd, 0x0591, 0x9504, 0x71cf, 0xffff, 0xedc0, 0x7915, 0x9106, 0x0d8a, 0x04a0, 0x9127, 0x9504, 0xe001,
+    0xd268, 0x7910, 0x094b, 0x0475, 0xb504, 0xd809, 0xf055, 0x087d, 0x0610, 0x08a1, 0x06d1, 0x72cf, 0xffff, 0xec80, 0x8a0e, 0x8a2f, 0xb88f, 0x1c00, 0x3004, 0x124a, 0x0100, 0xb908, 0x7905, 0x1c02, 0x3044,
+    0xd298, 0x2840, 0x0201, 0x7825, 0x1c04, 0x3004, 0x718b, 0x70cf, 0x0000, 0x3f72, 0x0df2, 0x04a0, 0xda06, 0xf01d, 0x70cf, 0xffff, 0xad34, 0x9002, 0xb8c2, 0xe083, 0xd818, 0x20ca, 0x05e2, 0xa501, 0xd800,
+    0xd2c8, 0xb504, 0xf1d4, 0x9504, 0x71cf, 0xffff, 0xee00, 0x7915, 0x9108, 0x0d0e, 0x04a0, 0x9129, 0x9504, 0xe001, 0x7910, 0x098d, 0x8934, 0xb504, 0xd800, 0xa501, 0xf1c2, 0x9504, 0x71cf, 0xffff, 0xee80,
+    0xd2f8, 0x7915, 0x9110, 0x0cea, 0x04a0, 0x9131, 0x9504, 0xe001, 0x7910, 0xe1ce, 0x0768, 0xffe5, 0xb504, 0xd81b, 0xf1da, 0xd800, 0xf003, 0xd80c, 0x0199, 0x1320, 0xc0a2, 0xc0f1, 0xc5e1, 0x71cf, 0xffff,
+    0xd328, 0xec80, 0xb804, 0x611d, 0x2156, 0x0580, 0x2556, 0x1381, 0x0ace, 0x11e0, 0xda10, 0x70cf, 0xffff, 0xed80, 0x2556, 0x1601, 0x0abe, 0x11e0, 0xda10, 0x0165, 0x1300, 0xc0f1, 0x08ce, 0x1300, 0x70cf,
+    0xd358, 0xffff, 0xacf4, 0x71cf, 0x0000, 0x2710, 0x099a, 0x1320, 0x8003, 0x240a, 0x2f80, 0xffff, 0xec80, 0x71cf, 0xffff, 0xecc0, 0x252f, 0x2008, 0xd801, 0x2159, 0x0203, 0xf003, 0xe001, 0x780f, 0x0817,
+    0xd388, 0x00d3, 0x2314, 0x0002, 0x1260, 0x0102, 0x0af3, 0x8541, 0xffe3, 0xf040, 0x9110, 0x70b2, 0xd800, 0xf67b, 0x9113, 0x70b2, 0x20ca, 0x00ec, 0xf6f7, 0xde00, 0xf004, 0xe601, 0x7ecf, 0x0e65, 0x10d3,
+    0xd3b8, 0x2114, 0x038f, 0x9710, 0x08f5, 0x8543, 0x9711, 0x0ded, 0xa003, 0xdd00, 0xf026, 0x6e14, 0x7082, 0x1720, 0x1113, 0x2014, 0x0351, 0x1722, 0x1110, 0x1170, 0x2100, 0x726a, 0x1180, 0x2101, 0x730a,
+    0xd3e8, 0x0e26, 0x1060, 0x240a, 0x0540, 0x2414, 0x2352, 0x1ab0, 0x2004, 0x11c0, 0x2100, 0x726a, 0x11d0, 0x2101, 0x730a, 0x0e0a, 0x1060, 0x240a, 0x0540, 0x1a80, 0x201c, 0xe501, 0x7daf, 0x0db9, 0x9212,
+    0xd418, 0x0059, 0x1300, 0xc0f1, 0x0812, 0x1300, 0x70cf, 0xffff, 0xd228, 0x0e12, 0x0480, 0x200a, 0x2f80, 0xffff, 0xacf4, 0x1038, 0x2002, 0x76cf, 0xffff, 0xeffc, 0x8621, 0x75cf, 0xffff, 0xad34, 0x210a,
+    0xd448, 0x2f80, 0x0000, 0x2000, 0x77cf, 0xffff, 0xec80, 0x0a23, 0x0041, 0x9520, 0x9604, 0x091b, 0x0001, 0x9502, 0x9626, 0x2004, 0x0440, 0xb84d, 0x080f, 0x0041, 0x9725, 0x9605, 0x0933, 0x0000, 0x70cf,
+    0xd478, 0x0002, 0x40c0, 0x0dbe, 0x0480, 0x70cf, 0x0001, 0x55b8, 0x0db6, 0x0480, 0x1038, 0x2000, 0xa601, 0x9500, 0xb604, 0x9705, 0xb605, 0x9502, 0x2004, 0x0440, 0xb84d, 0xb606, 0x70cf, 0xffff, 0xce90,
+    0xd4a8, 0x0d92, 0x0480, 0x70cf, 0x0001, 0x725c, 0x0d8a, 0x0480, 0x70cf, 0xffff, 0xcf68, 0x0d7e, 0x0480, 0x70cf, 0xffff, 0xcfd0, 0x0d76, 0x0480, 0x70cf, 0xffff, 0xd038, 0x0d6a, 0x0480, 0x70cf, 0xffff,
+    0xd4d8, 0xd084, 0x0d62, 0x0480, 0x70cf, 0xffff, 0xce60, 0x0d56, 0x0480, 0x70cf, 0xffff, 0xd160, 0x0d4e, 0x0480, 0x9502, 0xb8c2, 0xe083, 0x0e58, 0xffc1, 0x8f0d, 0x0795, 0x12e0, 0x1f5f, 0x1002, 0x78e0,
+    0xd508, 0xc5e1, 0xc6e1, 0x72cf, 0xffff, 0xa168, 0x1224, 0x0082, 0x0a23, 0x00d1, 0xdb01, 0xf003, 0xe301, 0x7b70, 0x0b47, 0x00d5, 0x72cf, 0xffff, 0xed80, 0x7a76, 0x925a, 0x08ef, 0x8083, 0xf019, 0x72cf,
+    0xd538, 0xffff, 0xad54, 0x8a75, 0x72cf, 0xffff, 0xed90, 0xe380, 0xdd02, 0x22c0, 0x0222, 0x25ca, 0x11e2, 0xdb01, 0xf004, 0xe301, 0x7b70, 0x0b0f, 0x0343, 0x2215, 0x00ce, 0x96c0, 0x08f5, 0x8383, 0xbb61,
+    0xd568, 0xb160, 0xc6c1, 0x7fe0, 0xc5c1, 0xc0f1, 0x71cf, 0xffff, 0xacf4, 0x1175, 0x0083, 0x72cf, 0xffff, 0xed80, 0x7a15, 0xeb83, 0x9208, 0xf007, 0x9134, 0xb925, 0x0f7e, 0x12e0, 0x920c, 0x7810, 0x07a9,
+    0xd598, 0xff8f, 0x78e0, 0x7ee0, 0x78e0, 0xc0f1, 0x0e96, 0x12c0, 0xc1a3, 0x76cf, 0xffff, 0xa168, 0x7508, 0xe98b, 0xd800, 0xb605, 0xd80d, 0xa601, 0x71cf, 0xffff, 0xb530, 0x70c9, 0x913a, 0xfe11, 0xe580,
+    0xd5c8, 0x0240, 0x0002, 0x8621, 0xe195, 0x0238, 0x000d, 0x2740, 0x7200, 0x7835, 0x7800, 0x0221, 0x0000, 0x0229, 0x0000, 0x0225, 0x0000, 0x0221, 0x0000, 0x021d, 0x0000, 0x0219, 0x0000, 0x0215, 0x0000,
+    0xd5f8, 0x0211, 0x0000, 0x020d, 0x0000, 0x0209, 0x0000, 0x0205, 0x0000, 0x0201, 0x0000, 0x01fd, 0x0000, 0x0025, 0x0000, 0x0041, 0x0000, 0x0059, 0x0000, 0x007f, 0x0000, 0x0097, 0x0000, 0x00af, 0x0000,
+    0xd628, 0x014b, 0x0000, 0x0177, 0x0000, 0x01a9, 0x0000, 0x1626, 0x1080, 0xe80e, 0x71cf, 0xffff, 0xecc0, 0x70cf, 0x0000, 0x3012, 0x09a2, 0x04a0, 0x9122, 0xd80e, 0xa601, 0xdd09, 0xf0dc, 0x1626, 0x1080,
+    0xd658, 0xe80c, 0x71cf, 0xffff, 0xecc0, 0x70cf, 0x0000, 0x321a, 0x0982, 0x04a0, 0x9123, 0xd80f, 0xf1f1, 0x162b, 0x1081, 0x70cf, 0xffff, 0xb530, 0xe911, 0x9003, 0x71cf, 0xffff, 0xecc0, 0x2082, 0x0c83,
+    0xd688, 0xb104, 0x70cf, 0x0000, 0x3226, 0x0956, 0x04a0, 0x9124, 0xd810, 0xf1db, 0x162a, 0x1080, 0xe80d, 0x71cf, 0xffff, 0xb570, 0x70cf, 0x0000, 0x300a, 0x093a, 0x04a0, 0x9120, 0xd811, 0xf1cd, 0x162b,
+    0xd6b8, 0x1080, 0xe80d, 0x71cf, 0xffff, 0xb570, 0x70cf, 0x0000, 0x300c, 0x091e, 0x04a0, 0x9121, 0xd812, 0xf1bf, 0x1627, 0x1080, 0x77cf, 0xffff, 0xb530, 0x0897, 0x0010, 0x75cf, 0xffff, 0xec80, 0x1557,
+    0xd6e8, 0x1081, 0x971d, 0x0925, 0x00d1, 0x0c86, 0x1320, 0x718b, 0x1400, 0x3101, 0x70cf, 0xffff, 0xed80, 0x7836, 0x903b, 0xd800, 0x1c02, 0x3044, 0x1d55, 0x1002, 0xf02c, 0x1745, 0x1081, 0xe916, 0x71cf,
+    0xd718, 0xffff, 0xacf4, 0x9134, 0x790c, 0x2942, 0x7100, 0xb83b, 0x7027, 0x0cfa, 0x1260, 0xb845, 0x72cf, 0x3fe0, 0x0000, 0x092a, 0x1220, 0xdb00, 0x08ce, 0x1280, 0x7810, 0x0c36, 0x1320, 0x2440, 0x3081,
+    0xd748, 0x1745, 0x1082, 0x70cf, 0xffff, 0xed80, 0x1402, 0x3101, 0x7835, 0xea03, 0x902d, 0xf002, 0x9029, 0x1c02, 0x3044, 0x70cf, 0x0000, 0x3366, 0x087e, 0x0480, 0xd813, 0xf16f, 0x1627, 0x1080, 0xe817,
+    0xd778, 0x70cf, 0xffff, 0xecc0, 0x8817, 0xe891, 0x70cf, 0xffff, 0xb570, 0x8805, 0xe080, 0x20ca, 0x0062, 0x7913, 0x70cf, 0x0000, 0x3362, 0x084e, 0x04a0, 0xb9c3, 0xd814, 0xf157, 0x1628, 0x1080, 0xe81a,
+    0xd7a8, 0x71cf, 0xffff, 0xb570, 0x9104, 0x1c04, 0x3004, 0x9106, 0x1c06, 0x3004, 0x9103, 0x1c08, 0x3004, 0x9105, 0xc181, 0x1c0a, 0x3004, 0x70cf, 0x0000, 0x3056, 0x08ce, 0x04a0, 0xda08, 0xd815, 0xf13d,
+    0xd7d8, 0x77cf, 0xffff, 0xecc0, 0x8f14, 0xe80e, 0x70cf, 0xffff, 0xb570, 0x883b, 0x70cf, 0x0000, 0x3308, 0x0ff6, 0x0460, 0xb903, 0xd800, 0xaf14, 0xf12b, 0x71cf, 0xffff, 0xb530, 0x9100, 0xb884, 0xb100,
+    0xd808, 0x70a9, 0x0499, 0x12e0, 0xc0a3, 0x7ee0, 0x78e0, 0xc0f1, 0xd808, 0x0bc2, 0x0fa0, 0x7108, 0x70cf, 0x0000, 0x301a, 0xd900, 0xb98b, 0x0842, 0x0b60, 0x7228, 0x70cf, 0xffff, 0xce60, 0x0a3e, 0x0480,
+    0xd838, 0x0505, 0xff8f, 0x7fe0, 0xd800, 0xc0f1, 0x0bf6, 0x12c0, 0x7708, 0x70cf, 0xffff, 0xaff4, 0x100b, 0x00cd, 0x73cf, 0xffff, 0xec80, 0x76cf, 0xffff, 0xece0, 0x1608, 0x10c1, 0x2659, 0x1300, 0x7530,
+    0xd868, 0x134a, 0x0102, 0xd900, 0xf692, 0x160e, 0x10c1, 0x7530, 0xd907, 0xf60c, 0xd901, 0xf003, 0xe101, 0x792f, 0x0911, 0x01d5, 0x603e, 0x1668, 0x10ce, 0x0df3, 0x9383, 0x70cf, 0xffff, 0xb530, 0x90bd,
+    0xd898, 0x70cf, 0x0000, 0x1388, 0x7b34, 0x0d0b, 0x1811, 0x13b0, 0x0101, 0xf003, 0x1380, 0x0701, 0x794c, 0x71cf, 0x0000, 0x2710, 0x0c46, 0x12e0, 0x7027, 0x7810, 0xd9ff, 0x2009, 0x0041, 0x7930, 0x7a3b,
+    0xd8c8, 0x70cf, 0xffff, 0xacf4, 0xb054, 0x2704, 0x1f80, 0x0000, 0xff00, 0x03c9, 0x12e0, 0x7825, 0x78e0, 0xc0f1, 0xc5e1, 0x75cf, 0xffff, 0xa168, 0xc1a1, 0xe98a, 0x71cf, 0xffff, 0xecc0, 0x8937, 0xe183,
+    0xd8f8, 0xd920, 0x21ca, 0x08e2, 0xa521, 0x08c9, 0x0011, 0x8521, 0xe1a0, 0xf22c, 0xf619, 0x09bd, 0x0010, 0x09b9, 0x06d1, 0x9505, 0xffcb, 0x8d2b, 0x1c00, 0x3004, 0x2940, 0x0200, 0x7825, 0x1c02, 0x3004,
+    0xd928, 0x718b, 0x70cf, 0x0000, 0x3f74, 0x0f6a, 0x0460, 0xda04, 0xd823, 0xf046, 0x097d, 0x0850, 0x098d, 0x08d1, 0x72cf, 0xffff, 0xb530, 0x9220, 0x0981, 0x025e, 0xd8b0, 0xb88d, 0x7148, 0x0ede, 0x0460,
+    0xd958, 0x2180, 0x0183, 0xd800, 0xf034, 0x70cf, 0xffff, 0xaff4, 0x100b, 0x00c2, 0x71cf, 0xffff, 0xeffc, 0x110e, 0x00c0, 0xe055, 0x0811, 0x0082, 0xda0e, 0xba8c, 0xd803, 0xa90e, 0xf012, 0x110f, 0x00c0,
+    0xd988, 0xe05f, 0x0813, 0x00a3, 0xdb00, 0x72cf, 0x0000, 0x0f3c, 0xa96f, 0xf006, 0xda7d, 0xba05, 0xa96e, 0xd803, 0xa90f, 0x70cf, 0x0000, 0x337a, 0x0e3e, 0x0460, 0x7148, 0xd821, 0xf00a, 0x70cf, 0x0000,
+    0xd9b8, 0x3f74, 0x0e7a, 0x0460, 0x2540, 0x1281, 0xd81b, 0xa501, 0xd809, 0x02e9, 0x12e0, 0xc0a1, 0x78e0, 0xc0f1, 0x0a62, 0x12c0, 0x76cf, 0xffff, 0xb530, 0x966a, 0x200a, 0x2f80, 0xffff, 0xa168, 0x9649,
+    0xd9e8, 0x6b32, 0x86ec, 0x2a05, 0x007e, 0x0f1b, 0x1e45, 0x2f41, 0x11c0, 0x7810, 0x2048, 0x0040, 0x2009, 0x0080, 0x7d10, 0xb60b, 0xbd07, 0xf017, 0x70cf, 0xffff, 0xecc0, 0x881d, 0xe88e, 0x2715, 0x10c0,
+    0xda18, 0x0ae2, 0x12e0, 0xb861, 0x7810, 0x2048, 0x0041, 0x9608, 0x2009, 0x0040, 0xf1ed, 0x86a3, 0x70a9, 0xe07f, 0xb827, 0xb60b, 0x6f17, 0x0ac2, 0x12e0, 0x71a9, 0xa60a, 0x9605, 0x2805, 0x037e, 0x2941,
+    0xda48, 0x71c0, 0xb60d, 0x101e, 0x2100, 0x0249, 0x12e0, 0xb60c, 0x78e0, 0xc0f1, 0x09d2, 0x12c0, 0x711a, 0x210a, 0x2f80, 0xffff, 0xec80, 0x1157, 0x2081, 0xe183, 0xc1a1, 0xf499, 0x76cf, 0xffff, 0xb530,
+    0xda78, 0x9630, 0x7822, 0xd908, 0x0846, 0x1060, 0xda07, 0xa60c, 0x9602, 0x081b, 0x001e, 0xffd1, 0xd801, 0x1ec4, 0x1002, 0x960c, 0x71cf, 0xffff, 0xad34, 0xb113, 0x960d, 0xb112, 0x9602, 0x08d5, 0x005e,
+    0xdaa8, 0x2840, 0x21c0, 0x75cf, 0xffff, 0xadb4, 0x0a5a, 0x12e0, 0x9523, 0x72cf, 0xffff, 0xb5b0, 0xb219, 0x9239, 0xd880, 0x2108, 0x0001, 0x921a, 0x220a, 0x2f80, 0xffff, 0xacf4, 0x2009, 0x0040, 0xb219,
+    0xdad8, 0x9239, 0x860a, 0x7030, 0x21ca, 0x0005, 0x0a1a, 0x12e0, 0xb805, 0x9525, 0x2905, 0x003e, 0x2941, 0x7140, 0x7110, 0x20ca, 0x004d, 0x9526, 0x7110, 0x20ca, 0x0045, 0x7810, 0x0876, 0x1320, 0x718b,
+    0xdb08, 0x1400, 0x3101, 0x70cf, 0xffff, 0xed80, 0x864a, 0x7836, 0x903a, 0x70cf, 0xffff, 0xad34, 0xb034, 0xb035, 0xb036, 0xb037, 0x9505, 0x2805, 0x00be, 0x09ce, 0x12e0, 0x702f, 0x95e3, 0xda40, 0x2f05,
+    0xdb38, 0x103e, 0x2941, 0x71c0, 0xb806, 0x2841, 0x01c1, 0xe1c0, 0x22ca, 0x004d, 0x114e, 0x2101, 0x7150, 0x21ca, 0x0086, 0x09a6, 0x12e0, 0x1a8b, 0x2042, 0x7710, 0x21ca, 0x000e, 0x9504, 0x21ca, 0x03cd,
+    0xdb68, 0x7030, 0x21ca, 0x0005, 0x1a86, 0x2044, 0xd901, 0x1a75, 0x2042, 0x9602, 0x0847, 0x009e, 0x70cf, 0xffff, 0xacf4, 0x10bc, 0x0080, 0x75cf, 0xffff, 0xad34, 0x0811, 0x005e, 0x0936, 0x0da0, 0x9512,
+    0xdb98, 0xb518, 0xf013, 0x9604, 0xb518, 0xf00f, 0x72cf, 0x0001, 0x8d02, 0x7a60, 0x710a, 0x208a, 0x0806, 0x71cf, 0xffff, 0xacf4, 0xb114, 0xd840, 0x198b, 0x0002, 0x00cd, 0x12e0, 0xc0a1, 0xc0f1, 0x0876,
+    0xdbc8, 0x12c0, 0x70cf, 0xffff, 0x8284, 0x2040, 0x080d, 0x9013, 0x8dc0, 0x0811, 0x00d1, 0xe680, 0x26cc, 0x9062, 0xf404, 0xd802, 0xad00, 0x70cf, 0x0001, 0x7e2a, 0x7840, 0xadc0, 0x73cf, 0xffff, 0xb570,
+    0xdbf8, 0x8bbb, 0x71cf, 0xffff, 0xacf4, 0x118b, 0x0082, 0x0d23, 0x1080, 0x812c, 0xab5b, 0xd804, 0x2104, 0x0f81, 0x0000, 0xe000, 0x0ec2, 0x0ce0, 0xb92d, 0xd801, 0x71cf, 0xffff, 0xecc0, 0xa914, 0x0085,
+    0xdc28, 0x12c0, 0x78e0, 0xc0f1, 0xc5e1, 0x75cf, 0xffff, 0xa168, 0x72cf, 0xffff, 0xefd8, 0x8248, 0xc2e1, 0xe983, 0xd933, 0xa521, 0xe8ac, 0x8521, 0xe1af, 0x0050, 0x000b, 0xf618, 0x9505, 0xe8a6, 0x208a,
+    0xdc58, 0x0508, 0x1c00, 0x3004, 0x208a, 0x0e06, 0x1c02, 0x3004, 0x718b, 0x70cf, 0x0000, 0x30c6, 0x0c2e, 0x0460, 0xda04, 0xd900, 0x70cf, 0xffff, 0xeffc, 0xa830, 0xf019, 0x0933, 0x0c90, 0x0919, 0x0cd1,
+    0xdc88, 0x70cf, 0x0000, 0x300e, 0x0ba6, 0x0460, 0x2540, 0x1281, 0xd82f, 0xa501, 0xd809, 0x052d, 0xffcf, 0x70cf, 0x0000, 0x30c6, 0x0b8e, 0x0460, 0x2540, 0x1281, 0xd832, 0xf1f4, 0x70cf, 0xffff, 0xeffc,
+    0xdcb8, 0x8810, 0x0811, 0x0051, 0x9505, 0xe007, 0xb806, 0x7910, 0xf005, 0x71cf, 0x0000, 0x8780, 0x70cf, 0x0000, 0x3f94, 0x0b12, 0x0440, 0xd800, 0xf1e0, 0xc0f1, 0x70cf, 0xffff, 0xd228, 0x0d8e, 0x0440,
+    0xdce8, 0xe886, 0x70cf, 0xffff, 0xdc2c, 0x0d82, 0x0440, 0x0049, 0xff8f, 0x70cf, 0x0000, 0x2510, 0x7ee0, 0xc0f1, 0x0f3a, 0x1280, 0x75cf, 0xffff, 0xa168, 0x7608, 0xe98f, 0x70cf, 0xffff, 0xb530, 0x71cf,
+    0xdd18, 0xffff, 0xf00e, 0x0e5a, 0x12e0, 0x901d, 0xd800, 0xb504, 0xd82b, 0xa501, 0xeea9, 0x8501, 0xe827, 0x0829, 0x0490, 0x0847, 0x0ad1, 0x70cf, 0xffff, 0xeffc, 0x9009, 0x71cf, 0xffff, 0xed80, 0x7916,
+    0xdd48, 0x70cf, 0x0000, 0x3f78, 0x0a9a, 0x0460, 0x913d, 0xd812, 0xf011, 0x70cf, 0xffff, 0xeffc, 0x9009, 0x71cf, 0xffff, 0xed80, 0x7916, 0x70cf, 0x0000, 0x3364, 0x0a7a, 0x0460, 0x913c, 0xd800, 0xa501,
+    0xdd78, 0xde09, 0x0731, 0x12a0, 0x70c9, 0xc0f1, 0xe080, 0x21cc, 0x8021, 0xf409, 0x70cf, 0x0000, 0x325e, 0x0a56, 0x0460, 0x218a, 0x0186, 0xd809, 0x07a5, 0xff4f, 0x78e0, 0xc0f1, 0xc5e1, 0x75cf, 0xffff,
+    0xdda8, 0xa168, 0xe894, 0xe983, 0xd939, 0xa521, 0x8521, 0xe90e, 0x091d, 0x0e51, 0x0f8a, 0xff4f, 0x7108, 0x70cf, 0x0000, 0x321e, 0x0a22, 0x0440, 0xd800, 0xa501, 0xd809, 0x06e1, 0x1280, 0xc0f1, 0x0e66,
+    0xddd8, 0x1280, 0x71cf, 0x0000, 0x552a, 0x7960, 0xdd00, 0x71cf, 0xffff, 0xacf4, 0x1146, 0x010e, 0x73cf, 0xffff, 0xec80, 0x72cf, 0xffff, 0xecc0, 0x78dd, 0xb8c0, 0xaa16, 0x70cf, 0xffff, 0xad74, 0x0e47,
+    0xde08, 0x105e, 0x90d8, 0x268c, 0x91cf, 0xaaa2, 0xf20a, 0x2602, 0x9f8d, 0xffff, 0xffeb, 0x25c0, 0x13e4, 0xbd44, 0x7daf, 0xaaa3, 0x90b7, 0x9018, 0xb2a8, 0xb209, 0x1144, 0x0100, 0x9322, 0xb8c2, 0xaa17,
+    0xde38, 0xd840, 0x2108, 0x0001, 0xd8ff, 0x2109, 0x0000, 0x0665, 0x12a0, 0xb207, 0x90d7, 0x2640, 0x938e, 0x26c0, 0x13e4, 0xbe44, 0xaac2, 0xf1e6, 0x78e0, 0xc0f1, 0x0dde, 0x1280, 0x76cf, 0xffff, 0xa188,
+    0xde68, 0x8e11, 0x0881, 0x0010, 0x75cf, 0xffff, 0xecc0, 0x8d19, 0xe80e, 0x0f8e, 0x0400, 0xe80a, 0x70cf, 0xffff, 0xdd80, 0x0bb6, 0x0440, 0x0fae, 0x0400, 0xd800, 0xad19, 0x8e07, 0xe810, 0x8d15, 0xe88e,
+    0xde98, 0x0f6e, 0x0400, 0xe80a, 0x70cf, 0xffff, 0xdd00, 0x0b96, 0x0440, 0x0f8e, 0x0400, 0xd801, 0xad15, 0x8e07, 0xe89a, 0x8e06, 0xe898, 0x950d, 0x8d3f, 0xe001, 0xb50d, 0x8d1e, 0x6038, 0x952d, 0x0925,
+    0xdec8, 0x0002, 0x0f3e, 0x0400, 0xe80e, 0x70cf, 0xffff, 0xdda0, 0x0b66, 0x0440, 0x0f5e, 0x0400, 0x0e4e, 0xff4f, 0xad1e, 0xd800, 0xb50d, 0x70cf, 0x0001, 0x8016, 0x7840, 0x05b9, 0x1280, 0x71cf, 0xffff,
+    0xdef8, 0xecc0, 0x891c, 0x72cf, 0xffff, 0xec80, 0xe885, 0xd801, 0xa919, 0xa91c, 0x8205, 0x7800, 0x78e0, 0xc0f1, 0x0d26, 0x1280, 0x75cf, 0xffff, 0xacf4, 0x951f, 0xc1a1, 0x0815, 0x0f80, 0x0000, 0x1d54,
+    0xdf28, 0x70cf, 0x0001, 0x42fe, 0x7840, 0x7608, 0xf0f6, 0x70cf, 0xffff, 0xc0bc, 0x71cf, 0xffff, 0xd5a0, 0xa020, 0x70cf, 0xffff, 0xc090, 0x71cf, 0xffff, 0xd41c, 0xa020, 0x70cf, 0xffff, 0xc08c, 0x71cf,
+    0xdf58, 0xffff, 0xd508, 0xa020, 0x70cf, 0xffff, 0xc080, 0x71cf, 0xffff, 0xd814, 0xa020, 0x70cf, 0xffff, 0xc078, 0x71cf, 0x0001, 0x5418, 0xa020, 0x70cf, 0xffff, 0xc0dc, 0x71cf, 0xffff, 0xd83c, 0xa020,
+    0xdf88, 0x70cf, 0xffff, 0xc084, 0x71cf, 0xffff, 0xd570, 0xa020, 0x70cf, 0xffff, 0xc0c4, 0x71cf, 0xffff, 0xd8e0, 0xa020, 0x70cf, 0xffff, 0xc1d8, 0x71cf, 0xffff, 0xd59c, 0xa020, 0x70cf, 0xffff, 0xc0d4,
+    0xdfb8, 0x71cf, 0xffff, 0xd160, 0xa020, 0x70cf, 0xffff, 0xc0e0, 0x71cf, 0xffff, 0xd810, 0xa020, 0x70cf, 0xffff, 0xc0f8, 0x71cf, 0xffff, 0xcfd0, 0xa020, 0x71cf, 0xffff, 0xc160, 0x8100, 0x76cf, 0xffff,
+    0xdfe8, 0xec80, 0xa605, 0x70cf, 0xffff, 0xdef4, 0xa100, 0x208a, 0x0015, 0xb500, 0x208a, 0x020f, 0x0bca, 0x0320, 0xb501, 0x73cf, 0xffff, 0xad74, 0x240a, 0x0f80, 0xffff, 0xb0f4, 0xda20, 0xe894, 0xb543,
+    0xe018, 0x208a, 0x07d4, 0xb505, 0xd87c, 0xb502, 0x208a, 0x02cd, 0xb504, 0x218a, 0x0014, 0xb332, 0x208a, 0x040b, 0xb313, 0x1c72, 0x805c, 0x1c73, 0x801c, 0x208a, 0x0204, 0x77cf, 0xffff, 0xad34, 0xb700,
+    0xe048, 0x218a, 0x021a, 0xb52b, 0x208a, 0x01d0, 0xb50a, 0x9702, 0xb529, 0x2004, 0x0f80, 0xffff, 0xd83f, 0xb702, 0x9702, 0xb528, 0x71cf, 0xffff, 0xadb4, 0xb145, 0xb88b, 0xb702, 0xd8a8, 0xb306, 0xdb80,
+    0xe078, 0xb163, 0x208a, 0x0010, 0xb104, 0xb54d, 0xb56e, 0xb166, 0xd903, 0x72cf, 0xffff, 0xae54, 0x8a1d, 0xb72a, 0xdb7d, 0xb884, 0xaa1d, 0x70cf, 0xffff, 0xb5f0, 0xa86c, 0x238a, 0x0988, 0xda00, 0xb067,
+    0xe0a8, 0x1c14, 0x8084, 0x238a, 0x0087, 0x72cf, 0xffff, 0xb530, 0xb068, 0x9201, 0xae2c, 0xb889, 0xb88b, 0xb201, 0xd804, 0xb604, 0xd8ff, 0xb602, 0xd878, 0xae0e, 0xd83a, 0xae0f, 0x70cf, 0x0000, 0x3f74,
+    0xe0d8, 0x0eb6, 0x0420, 0x718b, 0x260a, 0x9000, 0xf41a, 0x1401, 0x3080, 0x208c, 0x8002, 0x71cf, 0xffff, 0xecc0, 0xb105, 0xf403, 0xd8e8, 0xb105, 0x9722, 0xb9c2, 0x090b, 0x00d1, 0x781b, 0xf003, 0x208a,
+    0xe108, 0x0806, 0xb514, 0xfef4, 0x260a, 0x9000, 0xf206, 0x70c9, 0x0bea, 0x0ca0, 0xd907, 0x70c9, 0x0385, 0x12a0, 0xc0a1, 0xc0f1, 0x0b06, 0x1280, 0x210a, 0x2f80, 0xffff, 0xb530, 0x11b2, 0x2101, 0x1128,
+    0xe138, 0x2000, 0xc1a3, 0x7030, 0x21ca, 0x0005, 0x0bba, 0x12a0, 0xb805, 0xdd20, 0x76cf, 0xffff, 0xadb4, 0x9645, 0x77cf, 0xffff, 0xacf4, 0x2a05, 0x003e, 0x111c, 0x2100, 0x2941, 0x7152, 0xd900, 0x081b,
+    0xe168, 0x04a5, 0xc142, 0x1c00, 0x3480, 0x708b, 0xc182, 0x0986, 0x0d60, 0xc281, 0xc102, 0xe983, 0x751a, 0xf003, 0x1728, 0x1110, 0x1f75, 0x1042, 0x75cf, 0xffff, 0xad34, 0x9522, 0x9606, 0x090d, 0x02de,
+    0xe198, 0x9734, 0xb925, 0x782c, 0x702f, 0x9625, 0x7152, 0x21ca, 0x048e, 0x7110, 0x20ca, 0x0045, 0x0b12, 0x0ca0, 0x7810, 0x2805, 0x043e, 0xb514, 0xb515, 0xb516, 0xb517, 0x1128, 0x2002, 0x9605, 0x2941,
+    0xe1c8, 0x7141, 0x2805, 0x00be, 0x0b2e, 0x12a0, 0x702f, 0x9623, 0x2905, 0x003e, 0x2941, 0x71c0, 0x7110, 0x20ca, 0x004d, 0x9624, 0x7110, 0x20ca, 0x0045, 0x1f86, 0x1004, 0x0299, 0x12a0, 0xc0a3, 0x78e0,
+    0xe1f8, 0xc0f1, 0x0a3e, 0x1280, 0x72cf, 0xffff, 0x82f0, 0x8a41, 0x0a0f, 0x01d1, 0x72cf, 0xffff, 0x859c, 0xf004, 0x72cf, 0xffff, 0xb960, 0x92a2, 0x72cf, 0xffff, 0xacf4, 0x92d4, 0xbd25, 0xdb00, 0xf019,
+    0xe228, 0x20f4, 0x00c2, 0x77cf, 0x0000, 0xffff, 0x7acc, 0x2904, 0x707e, 0x2941, 0x7382, 0x72f1, 0x27ca, 0x1085, 0x72cf, 0xffff, 0xec80, 0x7a75, 0x7ff0, 0x7fac, 0xe301, 0x7b6f, 0x1a18, 0x0e40, 0x0bd5,
+    0xe258, 0x80d2, 0xda00, 0xf015, 0x20f4, 0x0083, 0x75cf, 0x0000, 0xffff, 0x7b2c, 0x2940, 0x7103, 0xbb2d, 0x73b1, 0x25ca, 0x10c5, 0x73cf, 0xffff, 0xec80, 0x7b55, 0xe201, 0x7db0, 0x7a4f, 0xa3a9, 0x0adb,
+    0xe288, 0x80d2, 0xdb00, 0xf015, 0x20f4, 0x00c2, 0x75cf, 0x0000, 0xffff, 0xba07, 0x7a2c, 0x2941, 0x7382, 0x72b1, 0x25ca, 0x1085, 0x72cf, 0xffff, 0xec80, 0x7a75, 0xe301, 0x7db0, 0x7b6f, 0xa2ac, 0x0bdb,
+    0xe2b8, 0x80d2, 0x01e9, 0x1280, 0x78e0, 0xc0f1, 0x0962, 0x1280, 0x7708, 0x711a, 0x71cf, 0xffff, 0xecc0, 0x9102, 0x723a, 0x72cf, 0xffff, 0xb530, 0x9243, 0xb861, 0x7a0c, 0x752f, 0x9103, 0x2580, 0x1c49,
+    0xe2e8, 0xb861, 0x91c4, 0x784c, 0x230a, 0x2f80, 0xffff, 0xacf4, 0x1328, 0x2100, 0x7627, 0x79bb, 0x2805, 0x037e, 0x79d5, 0x2e40, 0x1094, 0x2941, 0x7140, 0x78b4, 0x2015, 0x0382, 0x783d, 0x2a40, 0x0152,
+    0xe318, 0x09e2, 0x12a0, 0x7042, 0xb700, 0x2515, 0x1381, 0x783d, 0x09d6, 0x12a0, 0x7042, 0xb701, 0x2214, 0x2380, 0x09ca, 0x12a0, 0x718a, 0xb702, 0x1358, 0x210f, 0x2f05, 0x13be, 0x2940, 0x7080, 0x09b6,
+    0xe348, 0x12a0, 0x71a9, 0x78f4, 0x1800, 0x2000, 0x135a, 0x210f, 0x2f05, 0x13be, 0x2940, 0x7080, 0x099e, 0x12a0, 0x71a9, 0x60f8, 0x0115, 0x12a0, 0x1900, 0x2000, 0x78e0, 0xc0f1, 0x089e, 0x1280, 0x703a,
+    0xe378, 0x7628, 0x7748, 0x1cec, 0xb0c8, 0x200a, 0x3100, 0x210a, 0x3140, 0x230a, 0x2180, 0x75e8, 0xc083, 0xc181, 0xc282, 0xffcb, 0xc101, 0x140c, 0x3100, 0xc202, 0x2805, 0x007e, 0x71cf, 0x000f, 0xffff,
+    0xe3a8, 0x2905, 0x737e, 0x2941, 0x72c0, 0x7030, 0x25ca, 0x2046, 0x25ca, 0x2005, 0x140e, 0x3100, 0x2805, 0x00be, 0x2905, 0x737e, 0x2941, 0x72c0, 0x7030, 0x24ca, 0x2046, 0x24ca, 0x2005, 0x70cf, 0xffff,
+    0xe3d8, 0xb334, 0x9001, 0xb8e6, 0x22ca, 0x2501, 0x26ca, 0x2541, 0xf20d, 0x70cf, 0xffff, 0xaff4, 0x1040, 0x0016, 0x1044, 0x0015, 0x1048, 0x0012, 0x104c, 0x0014, 0x70cf, 0xffff, 0xec80, 0x9003, 0xd908,
+    0xe408, 0x0fba, 0x0fe0, 0x7228, 0x202f, 0x2008, 0x140e, 0x3100, 0xd905, 0x0faa, 0x0fe0, 0x7228, 0x272f, 0x2008, 0x1410, 0x3100, 0xd905, 0x0f9a, 0x0fe0, 0x7228, 0x242f, 0x0008, 0x8600, 0x72cf, 0xffff,
+    0xe438, 0xe001, 0x2805, 0x04fe, 0x2905, 0x737e, 0x2102, 0x7080, 0xb82d, 0x7910, 0x8601, 0x793d, 0x2805, 0x04fe, 0x2905, 0x737e, 0x2102, 0x7080, 0xb82d, 0x7b10, 0x8602, 0x2140, 0x07c6, 0x2805, 0x04fe,
+    0xe468, 0x2905, 0x737e, 0x2102, 0x7080, 0xb82d, 0x7810, 0x140e, 0x3101, 0x2e41, 0x0146, 0x228a, 0x0070, 0x2905, 0x00fe, 0x212f, 0x0e42, 0x7942, 0x1410, 0x3103, 0xb92a, 0x87a0, 0x2b05, 0x003e, 0x202f,
+    0xe498, 0x0e42, 0x7842, 0xc300, 0xb82a, 0x2d05, 0x10fe, 0x2140, 0x77c3, 0x87a1, 0xbb25, 0x2d05, 0x163e, 0x2905, 0x743e, 0x2905, 0x75fe, 0x2100, 0x7f85, 0x0003, 0xffff, 0x87a2, 0x2d41, 0x0485, 0x2d05,
+    0xe4c8, 0x167e, 0x2905, 0x713e, 0x2102, 0x708d, 0x144f, 0x308e, 0xbd2a, 0x6e54, 0x7222, 0x1a00, 0x0580, 0x1a04, 0x0540, 0x1a08, 0x0480, 0x1a0c, 0x0500, 0x6e52, 0x7a55, 0x7222, 0x1a30, 0x0184, 0xb279,
+    0xe4f8, 0xb23a, 0x1a36, 0x0144, 0xb21c, 0xb2bd, 0xd800, 0xb21e, 0xb21f, 0x1a40, 0x0004, 0x1a42, 0x0004, 0x0741, 0x1260, 0xc0a5, 0x78e0, 0xc0f1, 0x0f0e, 0x1240, 0x707a, 0x70cf, 0xffff, 0x82f0, 0x8801,
+    0xe528, 0xc1a9, 0x080f, 0x01d1, 0x72cf, 0xffff, 0x859c, 0xf004, 0x72cf, 0xffff, 0xb960, 0x71cf, 0xffff, 0xaf34, 0xd801, 0xf003, 0xe001, 0x780f, 0x0813, 0x00d3, 0x2116, 0x0003, 0x93bc, 0x9262, 0x0bf3,
+    0xe558, 0x8343, 0xb861, 0x780f, 0x7916, 0x911d, 0x1227, 0x0092, 0xc041, 0x911e, 0x77cf, 0xffff, 0xec80, 0x71cf, 0xffff, 0xaff4, 0xc044, 0x9110, 0xc042, 0x9111, 0xc045, 0x9118, 0xc043, 0x9119, 0xc046,
+    0xe588, 0x9209, 0x1c1c, 0x3004, 0x920a, 0x1c1e, 0x3004, 0x920c, 0x1c20, 0x3004, 0xc087, 0x714a, 0xff17, 0xdd00, 0xf029, 0x2715, 0x134e, 0x8606, 0xd905, 0x0e16, 0x0fe0, 0x7228, 0x202f, 0x2008, 0x8609,
+    0xe5b8, 0xd905, 0x0e0a, 0x0fe0, 0x7228, 0x212f, 0x2008, 0x860c, 0xd905, 0x0dfa, 0x0fe0, 0x7228, 0x252f, 0x0008, 0xc181, 0xc284, 0x78bb, 0x7083, 0x101c, 0x0106, 0xc540, 0x706a, 0x730a, 0x240a, 0x0440,
+    0xe5e8, 0x270a, 0x0480, 0xff61, 0xe501, 0x7daf, 0x0db5, 0x90d2, 0x8726, 0x706a, 0x8749, 0x0d02, 0x0720, 0x876c, 0x0681, 0x1260, 0xc0a9, 0xc0f1, 0xc5e1, 0x75cf, 0xffff, 0xb334, 0x9522, 0x70cf, 0xffff,
+    0xe618, 0x8968, 0x0929, 0x019e, 0x9521, 0x0921, 0x001e, 0x71cf, 0xffff, 0xecc0, 0x8937, 0xe008, 0x090b, 0x00d1, 0xffba, 0xf003, 0x0d36, 0x0700, 0x9500, 0xb880, 0xf004, 0x9500, 0xb8a0, 0x066d, 0x1260,
+    0xe648, 0xb500, 0x78e0, 0xc0f1, 0x0dee, 0x1240, 0x70cf, 0xffff, 0x82f0, 0x8801, 0x080f, 0x01d1, 0x75cf, 0xffff, 0x859c, 0xf004, 0x75cf, 0xffff, 0xb938, 0x95ca, 0x1527, 0x1080, 0x71cf, 0x0000, 0xffff,
+    0xe678, 0x78cc, 0x2942, 0x7140, 0xb83a, 0x7027, 0xb846, 0x2009, 0x0040, 0xb50a, 0x70cf, 0x0000, 0xceb6, 0x7840, 0x0619, 0x1260, 0xb5ca, 0xc0f1, 0x0d82, 0x1240, 0x70cf, 0xffff, 0x82f0, 0x8801, 0x080f,
+    0xe6a8, 0x01d1, 0x75cf, 0xffff, 0x859c, 0xf004, 0x75cf, 0xffff, 0xb960, 0x080d, 0x01d1, 0x76cf, 0xffff, 0x859c, 0xf005, 0x76cf, 0xffff, 0xb938, 0x1614, 0x1116, 0x1616, 0x1110, 0x1612, 0x1111, 0x1618,
+    0xe6d8, 0x1112, 0x95ea, 0x1516, 0x1113, 0x1527, 0x1082, 0x1518, 0x1115, 0x70cf, 0x0000, 0xffff, 0x7aec, 0x2942, 0x7141, 0xb93a, 0x7127, 0xb946, 0x2109, 0x0001, 0xb52a, 0x952b, 0x1512, 0x1114, 0x7a2c,
+    0xe708, 0x2942, 0x7141, 0xb93a, 0x7127, 0xb946, 0x2109, 0x0001, 0xb52b, 0x9529, 0x7a2c, 0x2942, 0x7141, 0xb93a, 0x7127, 0xb946, 0x2109, 0x0001, 0xb529, 0x952c, 0x7a2c, 0x1627, 0x1082, 0x2942, 0x7141,
+    0xe738, 0xb93a, 0x7127, 0xb946, 0x2109, 0x0001, 0xb52c, 0x962a, 0x7a2c, 0x2942, 0x7141, 0xb93a, 0x7127, 0xb946, 0x2109, 0x0001, 0xb62a, 0x962b, 0x7a2c, 0x2942, 0x7141, 0xb93a, 0x7127, 0xb946, 0x2109,
+    0xe768, 0x0001, 0xb62b, 0x9629, 0x7a2c, 0x2942, 0x7141, 0xb93a, 0x7127, 0xb946, 0x2109, 0x0001, 0xb629, 0x962c, 0x7a2c, 0x2942, 0x7141, 0xb93a, 0x7127, 0xb946, 0x2109, 0x0000, 0xb60c, 0x70cf, 0x0000,
+    0xe798, 0xc43a, 0x7840, 0xb5ea, 0x1d16, 0x14c4, 0x1d12, 0x1504, 0x1d18, 0x1544, 0x1e14, 0x1584, 0x1e16, 0x1404, 0x1e12, 0x1444, 0x04b5, 0x1260, 0x1e18, 0x1484, 0x78e0, 0xc0f1, 0x0c7a, 0x1240, 0x76cf,
+    0xe7c8, 0xffff, 0xad34, 0x9602, 0x71cf, 0xffff, 0xa168, 0x1124, 0x0081, 0x2004, 0x0f8d, 0x0000, 0x0800, 0x0909, 0x00d1, 0xb8ab, 0xb602, 0x0d52, 0x0cc0, 0x9622, 0x7d25, 0x71cf, 0xffff, 0xecc0, 0x9127,
+    0xe7f8, 0xb6a2, 0x2905, 0x003e, 0x04ad, 0x1260, 0x2941, 0x7180, 0x78e0, 0xc0f1, 0x0c32, 0x1240, 0x75cf, 0xffff, 0xa168, 0x70cf, 0x0001, 0x85ae, 0x7840, 0x1524, 0x1080, 0x087d, 0x00d1, 0xffe7, 0x76cf,
+    0xe828, 0xffff, 0xb530, 0x7508, 0x9604, 0x71cf, 0xffff, 0xec80, 0x913f, 0xb86c, 0x7822, 0x7810, 0xb807, 0x0cca, 0x1260, 0x9625, 0x72cf, 0xffff, 0xacf4, 0x12bc, 0x0081, 0x2186, 0x0eff, 0x2942, 0x80be,
+    0xe858, 0xa603, 0xf20d, 0x12d2, 0x0102, 0x16b8, 0x1101, 0x0a0f, 0x0041, 0x2805, 0x037e, 0x2941, 0x71c0, 0xf005, 0x9608, 0x2805, 0x037e, 0x702f, 0x9630, 0xb62f, 0x0aa6, 0x0fe0, 0xd908, 0x9630, 0x2082,
+    0xe888, 0x001c, 0x6038, 0xb612, 0x6d15, 0x962a, 0xb827, 0x7110, 0x20ca, 0x0045, 0xb60a, 0x040d, 0x1240, 0xc0f1, 0x0b86, 0x1240, 0x0c2a, 0x02c0, 0x220a, 0x2f80, 0xffff, 0xacf4, 0x7508, 0x76cf, 0xffff,
+    0xe8b8, 0xb530, 0x12d1, 0x2080, 0x9623, 0x0a4a, 0x0ce0, 0x72a9, 0xb605, 0x9624, 0x9603, 0x200a, 0x2f80, 0xffff, 0xa168, 0x1022, 0x210f, 0x790c, 0x71a9, 0xd800, 0x7727, 0x0e66, 0x1260, 0xda08, 0x73e9,
+    0xe8e8, 0x0d62, 0x1260, 0xda00, 0x0a3a, 0x0ce0, 0x1ad2, 0x2044, 0x71cf, 0xffff, 0xb5b0, 0xb11c, 0x1024, 0x2080, 0x717a, 0x96a5, 0x9624, 0x089f, 0x00d1, 0x210a, 0x2f80, 0xffff, 0xec80, 0x113e, 0x2100,
+    0xe918, 0xb96c, 0x2102, 0x000f, 0x7ff0, 0x70e9, 0x0bea, 0x1260, 0x71a9, 0x202f, 0x2007, 0x6f17, 0x0bde, 0x1260, 0x71a9, 0x12bc, 0x2081, 0x2186, 0x0eff, 0x2942, 0x80be, 0xa603, 0xf208, 0x2805, 0x037e,
+    0xe948, 0x2941, 0x71cf, 0x7ff0, 0xf017, 0x1244, 0x2100, 0x2004, 0x0f80, 0x0000, 0x4000, 0x2842, 0x837e, 0xf209, 0x1108, 0x2100, 0x2048, 0x0041, 0x2109, 0x0400, 0x780f, 0xf002, 0x700a, 0x7d0c, 0x272f,
+    0xe978, 0x1e48, 0x70e9, 0x0b8e, 0x1260, 0xd950, 0x110c, 0x2081, 0x7810, 0x2148, 0x0041, 0x2109, 0x0001, 0x1958, 0x2042, 0x2702, 0x1001, 0x9603, 0x1b2e, 0x2044, 0x7d0c, 0x1b28, 0x2e40, 0xf00e, 0xb961,
+    0xe9a8, 0x7f30, 0x1b2e, 0x2044, 0x101e, 0x2101, 0x6f17, 0x1b28, 0x2040, 0x0b52, 0x1260, 0x71a9, 0xa603, 0x1b2c, 0x23c4, 0x70e9, 0x0b46, 0x1260, 0x71a9, 0x2048, 0x0040, 0xb609, 0xb608, 0x12bc, 0x2080,
+    0xe9d8, 0x71cf, 0xffff, 0xecc0, 0x2086, 0x0eff, 0xb842, 0x029d, 0x1260, 0xa91d, 0x78e0, 0xc0f1, 0x0a4e, 0x1240, 0x75cf, 0xffff, 0xa168, 0xe983, 0xd938, 0xa521, 0x088d, 0x0011, 0x8521, 0xe1b4, 0xf223,
+    0xea08, 0xf60e, 0x096d, 0x01d1, 0x9525, 0x0965, 0x005f, 0x2540, 0x1281, 0xd808, 0x0e1a, 0x03e0, 0xb88d, 0xd809, 0xf034, 0xde00, 0x0939, 0x0d50, 0x094d, 0x0e11, 0x70cf, 0xffff, 0xecc0, 0xa8dc, 0x70cf,
+    0xea38, 0x0000, 0x325e, 0x0daa, 0x03e0, 0x71c9, 0xd834, 0xa501, 0xd809, 0xf020, 0x70cf, 0x0000, 0x301a, 0x0de2, 0x03e0, 0x2540, 0x1281, 0xd835, 0xf1f6, 0x9525, 0x70cf, 0x0000, 0x301a, 0x2186, 0x0104,
+    0xea68, 0x0d7e, 0x03e0, 0xb525, 0xb5c5, 0xd807, 0xf1ea, 0xd900, 0x1d31, 0x1042, 0x72cf, 0xffff, 0xb530, 0x9220, 0x2186, 0x0202, 0xb220, 0x0221, 0x1240, 0xc0f1, 0xc5e1, 0x75cf, 0xffff, 0xeffc, 0xe984,
+    0xea98, 0xda37, 0xad54, 0x8d54, 0xea14, 0x0a25, 0x0dd1, 0x72cf, 0x0001, 0x6dc6, 0x7a40, 0xe88b, 0x0a96, 0xff0f, 0x7108, 0x70cf, 0x0000, 0x321e, 0x0d2e, 0x03c0, 0xd800, 0xad14, 0xd809, 0x01ed, 0x1240,
+    0xeac8, 0xd808, 0x7fe0, 0xb88d, 0x78e0, 0xc0f1, 0x096a, 0x1240, 0x7608, 0x72cf, 0xffff, 0xb530, 0x9200, 0x75cf, 0xffff, 0xacf4, 0xb8ab, 0x7905, 0xb220, 0x0a12, 0x03e0, 0x70c9, 0xd803, 0xb80c, 0x0c9a,
+    0xeaf8, 0x03e0, 0x2540, 0x1f81, 0xe884, 0x1d3c, 0x1382, 0x01a5, 0x1240, 0xc0f1, 0x0922, 0x1260, 0xd810, 0xc1a9, 0x0bb2, 0x0ee0, 0x7108, 0x09da, 0x0f20, 0xdd00, 0xd808, 0x08ba, 0x0f20, 0x7108, 0x0d4a,
+    0xeb28, 0x0760, 0x208a, 0x0004, 0x09b6, 0x0f00, 0x77cf, 0xffff, 0xad14, 0x8f16, 0x76cf, 0xffff, 0xacf4, 0x0d2e, 0x0760, 0xb808, 0xd80b, 0xf00a, 0x8f14, 0xd900, 0xffe1, 0xe805, 0x8f15, 0xd900, 0xb98b,
+    0xeb58, 0xffde, 0xe501, 0x228a, 0x2a0f, 0x0d07, 0x1483, 0xe8f3, 0x08df, 0x0011, 0x961f, 0x08d1, 0x0f81, 0x0000, 0x1d54, 0x70cf, 0xffff, 0xb530, 0x9021, 0x703a, 0xdd00, 0xb98a, 0xb021, 0x76a9, 0xf003,
+    0xeb88, 0xe601, 0x0e3b, 0x1483, 0x200a, 0x2700, 0xd80e, 0xb88d, 0x0bfa, 0x03e0, 0x710a, 0x250a, 0x9000, 0xf5f4, 0x1400, 0x3100, 0x08e5, 0x815e, 0xd80e, 0xb88d, 0x0be2, 0x03e0, 0x710a, 0x250a, 0x9000,
+    0xebb8, 0xf5e8, 0x1400, 0x3100, 0x08cd, 0x815e, 0xed84, 0x0e11, 0x1482, 0xd801, 0x0936, 0x0c60, 0xd923, 0x7508, 0xf008, 0xee06, 0x1100, 0x2100, 0xb88d, 0x1900, 0x2004, 0xedb6, 0x70cf, 0x0000, 0x301a,
+    0xebe8, 0xd9d0, 0x0b4e, 0x03e0, 0xb98c, 0x70cf, 0x0000, 0x31fe, 0x0b9a, 0x03e0, 0x718b, 0x1400, 0x3100, 0xc681, 0x71cf, 0xffff, 0xefd8, 0xb804, 0x2086, 0x03fc, 0xaf1d, 0x70c9, 0x09f2, 0x1120, 0xda20,
+    0xec18, 0x8f1c, 0x1c05, 0x3002, 0x0a3a, 0x03e0, 0x70c9, 0x0a7e, 0x03c0, 0xe8fe, 0x70cf, 0xffff, 0xb1bc, 0x2440, 0x3441, 0x09ce, 0x1120, 0xda10, 0xf00a, 0xd801, 0xd922, 0xf003, 0xd80b, 0xd901, 0x08ba,
+    0xec48, 0x0c40, 0x7508, 0x70a9, 0x003d, 0x1260, 0xc0a9, 0xd900, 0xf00a, 0x70cf, 0xffff, 0xf014, 0x7835, 0x8041, 0x8000, 0xe102, 0xa040, 0x09f1, 0x8794, 0x71cf, 0xffff, 0xec80, 0x70cf, 0xffff, 0xab9c,
+    0xec78, 0xb031, 0x7fe0, 0xd800, 0x0000, 0x0000, 0x0000, 0x00ff, 0x0012, 0x0004, 0x0000, 0x0364, 0x783a, 0x0000, 0x0011, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000,
+    0xeca8, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0007, 0x0001, 0x0008, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0001, 0x0000,
+    0xecd8, 0x0001, 0x0000, 0x0000, 0x1e00, 0x1388, 0x1450, 0x1518, 0x15e0, 0x1421, 0x303e, 0x4a5a, 0x6e00, 0x2694, 0x26a1, 0x26d9, 0x2710, 0x2740, 0x2770, 0x27b2, 0x27de, 0x26b9, 0x26c6, 0x26fd, 0x2734,
+    0xed08, 0x2768, 0x279b, 0x27e0, 0x2813, 0x26d0, 0x26dd, 0x2719, 0x2756, 0x278a, 0x27bd, 0x280f, 0x284f, 0x26e9, 0x26f5, 0x2734, 0x2773, 0x27ab, 0x27e3, 0x283d, 0x288d, 0x0000, 0x0000, 0x0000, 0x0000,
+    0xed38, 0x0000, 0x0000, 0x0000, 0x0000, 0x2635, 0x2661, 0x26b9, 0x2710, 0x273b, 0x2765, 0x27e2, 0x280b, 0x2661, 0x268e, 0x26b9, 0x2710, 0x2765, 0x278f, 0x280b, 0x285c, 0x268e, 0x26b9, 0x26e5, 0x2765,
+    0xed68, 0x278f, 0x27e2, 0x2834, 0x2884, 0x268e, 0x26b9, 0x273b, 0x27b9, 0x280b, 0x280b, 0x28fa, 0x296d, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0070, 0xcccc, 0x0080, 0xdddd,
+    0xed98, 0x01a0, 0x7777, 0x0208, 0x8888, 0x0270, 0x9999, 0x0340, 0xaaaa, 0x03c2, 0xbbbb, 0x05b0, 0xcccc, 0x0680, 0xdddd, 0x0020, 0xde79, 0x0004, 0x0210, 0x0040, 0xdeab, 0x0004, 0x0210, 0x0080, 0xdedc,
+    0xedc8, 0x0008, 0x0108, 0x3088, 0x0400, 0x3092, 0x1a24, 0x30b4, 0x0007, 0x3518, 0x14fe, 0x351a, 0x6000, 0x3526, 0x0f00, 0x352a, 0x089f, 0x352c, 0x0012, 0x353a, 0x9000, 0x353c, 0x7f00, 0x3542, 0x4b4b,
+    0xedf8, 0x3556, 0x1010, 0x3566, 0x7328, 0x30b0, 0x982e, 0x3524, 0x0c00, 0x31ae, 0x0001, 0x3116, 0x0001, 0x31d0, 0x0000, 0x31ac, 0x0c0c, 0x31ac, 0x0c0c, 0x3082, 0x0000, 0x30ba, 0x0020, 0x3372, 0x700f,
+    0xee28, 0x3014, 0x060e, 0x321e, 0x060e, 0x3222, 0x060e, 0x33e0, 0x0c80, 0x33e4, 0x0080, 0x33ee, 0x0344, 0x3364, 0x01cf, 0x32ea, 0x3c0e, 0x32ec, 0x72a1, 0x33e2, 0x0000, 0x322a, 0x0039, 0x3236, 0x00b2,
+    0xee58, 0x3c06, 0x083c, 0x3c08, 0x0100, 0x3180, 0x0080, 0x3100, 0x4000, 0x3102, 0x60a0, 0x3104, 0x60a0, 0x3106, 0x60a0, 0x3108, 0x0f9f, 0x3308, 0x0200, 0x3250, 0x0004, 0x3252, 0x0103, 0x3254, 0x2c3b,
+    0xee88, 0x3256, 0x0003, 0x3258, 0x0380, 0x3f72, 0x8080, 0x3f74, 0x40e1, 0x3f76, 0x8080, 0x3f78, 0x8080, 0x3754, 0x0000, 0x31d0, 0x0001, 0x3116, 0x0001, 0x31ac, 0x140c, 0x31ac, 0x140c, 0x3082, 0x000c,
+    0xeeb8, 0x30ba, 0x1023, 0x3528, 0xdddd, 0x352e, 0x00ff, 0x3530, 0xff00, 0x3532, 0x8f08, 0x3536, 0xff38, 0x3538, 0x24ff, 0x353c, 0x7f00, 0x3540, 0xc62c, 0x3014, 0x0271, 0x3222, 0x0879, 0x3226, 0x0594,
+    0xeee8, 0x33e0, 0x0080, 0x33e4, 0x0080, 0x33ee, 0x0344, 0x32ea, 0x3c69, 0x32ec, 0x7107, 0x33e2, 0x0530, 0x322a, 0x0939, 0x1008, 0x0271, 0x3230, 0x020f, 0x3236, 0x00c7, 0x3c06, 0x141c, 0x3c08, 0x2114,
+    0xef18, 0x30fe, 0x0040, 0x3180, 0x0188, 0x3504, 0x9100, 0x350c, 0x034a, 0x350e, 0x051c, 0x3510, 0x9753, 0x3520, 0x0800, 0x3522, 0x0008, 0x353e, 0x801f, 0x3544, 0x4b46, 0x354a, 0x007f, 0x3552, 0x0eb0,
+    0xef48, 0x3554, 0x1599, 0x355a, 0x0b0e, 0x3100, 0xc003, 0x3c82, 0x4fff, 0x3120, 0x0ba0, 0x3122, 0x0fa0, 0x312a, 0x83e8, 0x3250, 0x0005, 0x3252, 0x0107, 0x3256, 0x03e8, 0x3258, 0x0300, 0x325a, 0x0a13,
+    0xef78, 0x3260, 0x00ff, 0x3262, 0x00ff, 0x32d0, 0x3005, 0x32d2, 0x200a, 0x32d4, 0x3505, 0x32d6, 0x3207, 0x32dc, 0x220c, 0x32e2, 0x3707, 0x32f6, 0x3a01, 0x32f8, 0x3c03, 0x337a, 0x100e, 0x3372, 0x750f,
+    0xefa8, 0x3238, 0x8777, 0x3096, 0x0080, 0x3752, 0x0000, 0x3100, 0xe007, 0x3102, 0x8fa0, 0x3104, 0x8fa0, 0x3106, 0x9400, 0x3108, 0x0cb1, 0x3f7a, 0x01d6, 0x3f7c, 0x00c8, 0x31e0, 0x0003, 0x3362, 0x0f02,
+    0xefd8, 0x0200, 0x0407, 0x0902, 0x34c0, 0x0005, 0x080d, 0x1000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0600, 0x0000, 0x0000, 0x0036, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000,
+    0xf008, 0x0000, 0x0303, 0x0100, 0x0000, 0x3700, 0x0000, 0xffff, 0xc1b8, 0xffff, 0xda58, 0xffff, 0xc164, 0xffff, 0xdbc4, 0xffff, 0xc060, 0xffff, 0xdcf8, 0xffff, 0xbc90, 0xffff, 0xddd4, 0xffff, 0xc174,
+    0xf038, 0xffff, 0xde5c, 0xffff, 0xc06c, 0xffff, 0xdf10, 0xffff, 0xc1b4, 0xffff, 0xe124, 0xffff, 0xbe2c, 0xffff, 0xe608, 0xffff, 0xbe34, 0xffff, 0xe64c, 0xffff, 0xbe1c, 0xffff, 0xe698, 0xffff, 0xc194,
+    0xf068, 0xffff, 0xe808, 0xffff, 0xc190, 0xffff, 0xe8a0, 0xffff, 0xc0f4, 0xffff, 0xe9ec, 0xffff, 0xc0f0, 0xffff, 0xea8c, 0xffff, 0xc058, 0xffff, 0xeb08
+};
+
+// AP0202AT-REV2_AR0147-REV3.ini line 838
+static const uint16_t patch_01d4_data[] = {
+    0xc750, 0xc0f1, 0x0cea, 0x1340, 0x75cf, 0xffff, 0xa1a8, 0x8dc4, 0x0e0b, 0x10d1, 0xd804, 0xad04, 0x70cf, 0x0001, 0x7c06, 0x7840, 0x0e0f, 0x1111, 0xd800, 0x0cee, 0x0760, 0xad04, 0x0531, 0x1340, 0x78e0,
+    0xc780, 0xd900, 0xf00a, 0x70cf, 0xffff, 0xc79c, 0x7835, 0x8041, 0x8000, 0xe102, 0xa040, 0x09f1, 0x8094, 0x7fe0, 0xd800, 0xffff, 0xc160, 0xffff, 0xc750
+};
+
+// AP0202AT-REV2_AR0147-REV3.ini line 876
+static const uint16_t patch_03d4_data[] = {
+    0xc7a4, 0xc0f1, 0x0c92, 0x1340, 0x75cf, 0xffff, 0xb51c, 0x71cf, 0xffff, 0xa168, 0x812c, 0x77cf, 0xffff, 0x9d44, 0xb9e4, 0xd906, 0x21ca, 0x0221, 0xad32, 0x71cf, 0x0001, 0x2896, 0x7940, 0x260a, 0x9000,
+    0xc7d4, 0xf41e, 0x0efe, 0x0b20, 0x8d12, 0x0e1e, 0x0ae0, 0xd801, 0x70cf, 0xffff, 0xad34, 0x9002, 0x0829, 0x031e, 0x71cf, 0xff00, 0x0100, 0x1916, 0x0f84, 0x0000, 0x8022, 0x9703, 0xb10c, 0x191a, 0x0f84,
+    0xc804, 0x0000, 0x8002, 0x8f09, 0xb88d, 0xb88e, 0xb10e, 0x0491, 0x1360, 0x70c9, 0x78e0, 0xd900, 0xf00a, 0x70cf, 0xffff, 0xc834, 0x7835, 0x8041, 0x8000, 0xe102, 0xa040, 0x09f1, 0x8094, 0x7fe0, 0xd800,
+    0xc834, 0xffff, 0xbfe4, 0xffff, 0xc7a4
+};
+
+// AP0202AT-REV2_AR0147-REV3.ini line 916
+static const uint16_t patch_11d4_data[] = {
+    0xc83c, 0xc0f1, 0x0be6, 0x1340, 0x701a, 0x70cf, 0xffff, 0xac50, 0x90dd, 0x70cf, 0xffff, 0xb530, 0x9001, 0x713a, 0x080f, 0x005e, 0x0c12, 0x03c0, 0x7508, 0xf002, 0xdd80, 0x70cf, 0xffff, 0xae54, 0x883d,
+    0xc86c, 0x719a, 0xb9e4, 0x22ca, 0x2382, 0x22ca, 0x2341, 0x702a, 0x2009, 0x0401, 0xd880, 0x2109, 0x0000, 0x232f, 0x2008, 0x2a40, 0x21c0, 0x0c6e, 0x1360, 0x716a, 0x7708, 0x704a, 0x2805, 0x043e, 0x702f,
+    0xc89c, 0x0c5e, 0x1360, 0x716a, 0x701a, 0x702a, 0x2a05, 0x203e, 0x702f, 0x0c4e, 0x1360, 0x716a, 0x73cf, 0xffff, 0xad34, 0x72cf, 0xffff, 0xad74, 0x0c1b, 0x211e, 0xb3fc, 0xb3fd, 0x1b36, 0x0404, 0xb31e,
+    0xc8cc, 0xb2a0, 0xb2a1, 0xb3bf, 0xb2a2, 0xf00a, 0xb2e0, 0xb2e1, 0x1b3e, 0x0404, 0xb202, 0xb3dc, 0xb3dd, 0xb3db, 0xb3de, 0x0391, 0x1340, 0x70cf, 0xffff, 0xc83c, 0x71cf, 0xffff, 0xc94a, 0x72cf, 0x0000,
+    0xc8fc, 0xffff, 0xc5e1, 0xf010, 0x1001, 0x048d, 0x2a42, 0x0203, 0xba08, 0x7ba7, 0x7b6f, 0x6bbc, 0x7d67, 0x7aa7, 0x6d75, 0x7a67, 0xbd0c, 0x7aa7, 0x7a50, 0x08e3, 0x8044, 0x70cf, 0xffff, 0xc94a, 0xb040,
+    0xc92c, 0xd900, 0xf00a, 0x70cf, 0xffff, 0xc94c, 0x7835, 0x8041, 0x8000, 0xe102, 0xa040, 0x09f1, 0x8094, 0xd800, 0x7fe0, 0xc5c1, 0x0000, 0xffff, 0xc1e8, 0xffff, 0xc83c
+};
+
+// AP0202AT-REV2_AR0147-REV3.ini line 1046
+static const uint16_t patch_21d4_data[] = {
+	0xcc30, 0xc0f1, 0xc5e1, 0xe180, 0xd901, 0x75cf, 0x0002, 0x420e, 0x7d60, 0x79c0, 0x0071, 0x1340, 0x78e0, 0xc0f1, 0x0fee, 0x1300, 0x72cf, 0x0000, 0x4022, 0x7a60, 0x7728, 0x71cf, 0xffff, 0xafd4, 0x892a,
+	0xcc60, 0x76cf, 0xffff, 0x8284, 0xe182, 0xd901, 0x79c0, 0x75cf, 0xffff, 0xa168, 0x0f47, 0x1a11, 0x0843, 0x0250, 0xe99f, 0x71cf, 0xffff, 0xacf4, 0x1146, 0x0101, 0x091b, 0x001e, 0x153b, 0x10c2, 0x73cf,
+	0xcc90, 0xffff, 0xad74, 0x93f9, 0x7f42, 0xb3f9, 0x9359, 0xb647, 0x091b, 0x005e, 0x72cf, 0xffff, 0xad74, 0x923a, 0x153c, 0x10c3, 0x7962, 0xb23a, 0x923a, 0xb628, 0x07ed, 0x1300, 0x78e0, 0xd900, 0xf00a,
+	0xccc0, 0x70cf, 0xffff, 0xccd8, 0x7835, 0x8041, 0x8000, 0xe102, 0xa040, 0x09f1, 0x8114, 0x7fe0, 0xd800, 0xffff, 0xc11c, 0xffff, 0xcc30, 0xffff, 0xbc20, 0xffff, 0xcc48
+};
+
+// AP0202AT-REV2_AR0147-REV3.ini line 1443
+static const uint16_t patch_37d4_data[] = {
+	0xf08c, 0xc0f1, 0x0b82, 0x1200, 0x70cf, 0xffff, 0x82f0, 0x8821, 0xe187, 0x70cf, 0x0000, 0x185e, 0xc1a2, 0xf2ec, 0x71cf, 0xffff, 0x8190, 0xc140, 0x71cf, 0xffff, 0xb62c, 0x91e5, 0xc100, 0x210a, 0x3f80,
+	0xf0bc, 0xffff, 0xb334, 0x76cf, 0xffff, 0xb3b4, 0x1638, 0x1115, 0x75cf, 0xffff, 0xae74, 0x1104, 0x0018, 0x1108, 0x0016, 0x110c, 0x0014, 0x7860, 0x150c, 0x1013, 0x248c, 0xafff, 0x200a, 0x2f80, 0xffff,
+	0xf0ec, 0xacf4, 0xbf43, 0xf404, 0x0361, 0x1220, 0xc0a2, 0x70cf, 0xffff, 0xb614, 0x9000, 0x08b9, 0x001f, 0x164c, 0x1011, 0x10f3, 0x2700, 0x1650, 0x1012, 0x70f1, 0x27ca, 0x23ce, 0x27ca, 0x200d, 0x961d,
+	0xf11c, 0x11c0, 0x3101, 0x2805, 0x047e, 0x2941, 0x7100, 0xc041, 0x961e, 0xb904, 0x2805, 0x04be, 0x2941, 0x7100, 0x7110, 0x27ca, 0x100e, 0x08d6, 0x0d20, 0x27ca, 0x104d, 0x70f1, 0x21ca, 0x03c6, 0x21ca,
+	0xf14c, 0x0005, 0x965f, 0xc001, 0x258c, 0xa004, 0x7842, 0x2048, 0x0000, 0x2009, 0x0043, 0x208a, 0x0004, 0x22ca, 0x0006, 0x22ca, 0x0545, 0x7842, 0x2805, 0x063e, 0x762f, 0x2a05, 0x00fe, 0x7627, 0x2805,
+	0xf17c, 0x05be, 0x2e41, 0x120f, 0x762f, 0x2a05, 0x007e, 0x7627, 0x2805, 0x053e, 0xbe28, 0x702f, 0x2a05, 0x05fe, 0x7027, 0xc100, 0x2841, 0x0214, 0xa1e1, 0xa1c2, 0x1400, 0x3004, 0x718a, 0x1c0c, 0x0040,
+	0xf1ac, 0x9524, 0x0923, 0x00de, 0x85d5, 0x85f4, 0xf00f, 0x70cf, 0xffff, 0xac00, 0x9805, 0x0845, 0x8010, 0x0cb2, 0x00c0, 0x082b, 0x8011, 0xf19c, 0xa5d5, 0xa5f4, 0xb9e8, 0xb828, 0xd908, 0xf20e, 0x08ea,
+	0xf1dc, 0x0fa0, 0xda00, 0x2056, 0x2dc4, 0x71e9, 0x72c9, 0x0b4e, 0x0120, 0x856b, 0x1538, 0x1010, 0xf01d, 0x08ce, 0x0fa0, 0x7228, 0x7108, 0x2602, 0x13c0, 0x0afa, 0x1220, 0xb80c, 0x701a, 0x9504, 0x0813,
+	0xf20c, 0x001e, 0x704a, 0x0d16, 0x0120, 0x712a, 0x150c, 0x1013, 0x9525, 0xb904, 0x7102, 0x0ada, 0x1220, 0x2b40, 0x2100, 0x701a, 0x1d34, 0x1500, 0x7702, 0xe781, 0xd801, 0x20ca, 0x03cd, 0x08ea, 0x0fa0,
+	0xf23c, 0xd90c, 0x7508, 0x7602, 0xe681, 0xd801, 0x20ca, 0x038d, 0x08da, 0x0fa0, 0xd90c, 0x78a2, 0x2048, 0x0041, 0xd800, 0x0aa2, 0x1220, 0xb898, 0x72cf, 0x0001, 0xffff, 0x7550, 0x21ca, 0x0086, 0x21ca,
+	0xf26c, 0x0345, 0x7050, 0x22ca, 0x0005, 0x08de, 0x04a0, 0x700a, 0xf13d, 0x7840, 0xf13b, 0xd900, 0xf00a, 0x70cf, 0xffff, 0xf29c, 0x7835, 0x8041, 0x8000, 0xe102, 0xa040, 0x09f1, 0x8094, 0x7fe0, 0xd800,
+	0xf29c, 0xffff, 0xbbac, 0xffff, 0xf08c
+};
+
+// AP0202AT-REV2_AR0147-REV3.ini line 1491
+static const uint16_t patch_39d4_data[] = {
+	0xf2a4, 0xc0f1, 0xc5e1, 0xe986, 0xdb01, 0x72cf, 0xffff, 0xf344, 0xaa60, 0xe8b8, 0x75cf, 0xffff, 0xf344, 0x8d40, 0xea34, 0x0a65, 0x0051, 0x0951, 0x0151, 0x73cf, 0xffff, 0xacf4, 0x1344, 0x0100, 0xd901,
+	0xf2d4, 0x2004, 0x0f80, 0x0000, 0x2000, 0x2842, 0x837e, 0x70cf, 0xffff, 0xec80, 0x21c2, 0x0041, 0x9005, 0xb90f, 0xb808, 0x2004, 0x0f82, 0x0000, 0x1f00, 0x830e, 0x2004, 0x0f80, 0x007e, 0x0000, 0xb831,
+	0xf304, 0x7845, 0x7905, 0x70cf, 0x0000, 0x302c, 0x0cda, 0x0380, 0xf009, 0x72cf, 0x0002, 0x40ca, 0x7a40, 0xe883, 0xd800, 0xad00, 0xd809, 0x018d, 0x1200, 0xd900, 0xf00a, 0x70cf, 0xffff, 0xf348, 0x7835,
+	0xf334, 0x8041, 0x8000, 0xe102, 0xa040, 0x09f1, 0x8094, 0x7fe0, 0xd800, 0x0100, 0x0000, 0xffff, 0xc10c, 0xffff, 0xf2a4
+};
+
+// AP0202AT-REV2_AR0147-REV3.ini line 755
+static const uint16_t ar0147_sequencer_data[] = {
+	0x0901,0x3350,0x2004,0x1420,0x1578,0x087B,0x24FF,0x24FF,
+	0x24EA,0x2410,0x2224,0x1015,0xD813,0x0214,0x0024,0xFF24,
+	0xFF24,0xEA23,0x2464,0x7A24,0x0405,0x2C40,0x0AFF,0x0A78,
+	0x3851,0x2A54,0x1440,0x0015,0x0804,0x0801,0x0408,0x2652,
+	0x0813,0xC810,0x0210,0x1611,0x8111,0x8910,0x5612,0x1009,
+	0x020D,0x0903,0x1402,0x15A8,0x1388,0x0938,0x1199,0x11D9,
+	0x091E,0x1214,0x10D6,0x0901,0x1210,0x1212,0x1210,0x11DD,
+	0x11D9,0x0901,0x1441,0x0904,0x1056,0x0811,0xDB09,0x0311,
+	0xFB11,0xBB12,0x1A12,0x1008,0x1250,0x0B10,0x7610,0xE614,
+	0x6109,0x0612,0x4012,0x6009,0x1C14,0x6009,0x1215,0xC813,
+	0xC808,0x1066,0x090B,0x1588,0x1388,0x0913,0x0C14,0x4009,
+	0x0310,0xE611,0xFB12,0x6212,0x6011,0xFF11,0xFB14,0x4109,
+	0x0210,0x6609,0x1211,0xBB12,0x6312,0x6014,0x0015,0x1811,
+	0xB812,0xA012,0x0010,0x2610,0x0013,0x0011,0x0030,0x5342,
+	0x1100,0x1002,0x1016,0x1101,0x1109,0x1056,0x1210,0x0D09,
+	0x0314,0x0214,0x4309,0x0514,0x4009,0x0115,0xC813,0xC809,
+	0x1A11,0x4909,0x0815,0x8813,0x8809,0x1B11,0x5909,0x0B12,
+	0x1409,0x0112,0x1010,0xD612,0x1212,0x1011,0x5D11,0x5910,
+	0x5609,0x0311,0x5B08,0x1441,0x0901,0x1440,0x090C,0x117B,
+	0x113B,0x121A,0x1210,0x0901,0x1250,0x10F6,0x10E6,0x1460,
+	0x0901,0x15A8,0x13A8,0x1240,0x1260,0x0924,0x1588,0x0901,
+	0x1066,0x0B08,0x1388,0x0925,0x0C09,0x0214,0x4009,0x0710,
+	0xE612,0x6212,0x6011,0x7F11,0x7B10,0x6609,0x0614,0x4109,
+	0x0114,0x4009,0x0D11,0x3B12,0x6312,0x6014,0x0015,0x1811,
+	0x3812,0xA012,0x0010,0x2610,0x0013,0x0011,0x0043,0x7A06,
+	0x0507,0x410E,0x0237,0x502C,0x4414,0x4000,0x1508,0x0408,
+	0x0104,0x0826,0x5508,0x13C8,0x1002,0x1016,0x1181,0x1189,
+	0x1056,0x1210,0x0902,0x0D09,0x0314,0x0215,0xA813,0xA814,
+	0x0309,0x0614,0x0209,0x1F15,0x8813,0x8809,0x0B11,0x9911,
+	0xD909,0x1E12,0x1409,0x0312,0x1012,0x1212,0x1011,0xDD11,
+	0xD909,0x0114,0x4009,0x0711,0xDB09,0x0311,0xFB11,0xBB12,
+	0x1A12,0x1009,0x0112,0x500B,0x1076,0x1066,0x1460,0x0901,
+	0x15C8,0x0901,0x1240,0x1260,0x0901,0x13C8,0x0956,0x1588,
+	0x0901,0x0C14,0x4009,0x0511,0xFB12,0x6212,0x6011,0xFF11,
+	0xFB09,0x1911,0xBB12,0x6312,0x6014,0x0015,0x1811,0xB812,
+	0xA012,0x0010,0x2610,0x0013,0x0011,0x0030,0x5345,0x1444,
+	0x1002,0x1016,0x1101,0x1109,0x1056,0x1210,0x0D09,0x0314,
+	0x0614,0x4709,0x0514,0x4409,0x0115,0x9813,0x9809,0x1A11,
+	0x4909,0x0815,0x8813,0x8809,0x1B11,0x5909,0x0B12,0x1409,
+	0x0112,0x1009,0x0112,0x1212,0x1011,0x5D11,0x5909,0x0511,
+	0x5B09,0x1311,0x7B11,0x3B12,0x1A12,0x1009,0x0112,0x5010,
+	0x7610,0x6614,0x6409,0x0115,0xA813,0xA812,0x4012,0x6009,
+	0x2015,0x8809,0x020B,0x0901,0x1388,0x0925,0x0C09,0x0214,
+	0x4409,0x0912,0x6212,0x6011,0x7F11,0x7B09,0x1C11,0x3B12,
+	0x6312,0x6014,0x0015,0x1811,0x3812,0xA012,0x0010,0x2610,
+	0x0013,0x0011,0x0008,0x7A06,0x0508,0x070E,0x0237,0x502C,
+	0xFE32,0xFE06,0x2C2C
+};
+
+#endif // (OMV_ENABLE_AP0202AT_AR0147 == 1)
